% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bowtie2.R
\name{bowtie2}
\alias{bowtie2}
\title{Running bowtie2}
\usage{
bowtie2(
  group = c("sudo", "docker"),
  fastq.folder = getwd(),
  scratch.folder = "/data/scratch",
  genome.folder,
  seq.type = c("se", "pe"),
  strandness = c("none", "forward", "reverse"),
  threads = 1
)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{fastq.folder, }{a character string indicating where gzip fastq files are located}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{genome.folder, }{a character string indicating the folder where the indexed reference genome for STAR is located. IMPORTANT the present function only suport genomic indexes made using ensembl genom and the corresponding gtf}

\item{seq.type, }{a character string indicating the type of reads to be trimmed. Two options: \code{"se"} or \code{"pe"} respectively for single end and pair end sequencing}

\item{strandness, }{a character string indicating the type ofsequencing protocol used for the analysis. Three options: \code{"none"}, \code{"forward"}, \code{"reverse"} respectively for non strand selection, forward for Illumina strandness protocols, reverse for ACCESS Illumina protocol}

\item{threads, }{a number indicating the number of cores to be used from the application}
}
\value{
sorted.bam, sorted.bam.bai
}
\description{
This function executes the docker container bowtie2
}
\examples{
\dontrun{
    #downloading fastq files
system("wget http://130.192.119.59/public/test_R1.fastq.gz")
system("wget http://130.192.119.59/public/test_R2.fastq.gz")
library(docker4seq)
#running bowtie nostrand pe
bowtie2(group="docker",fastq.folder=getwd(), scratch.folder="/data/scratch/",
         genome.folder="/data/genomes/hg38bowtie2/", seq.type="pe", strandness="none",
         threads=8)

}
}
\author{
Raffaele Calogero
}
