% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script_bwaIndex.R
\name{bwaIndex}
\alias{bwaIndex}
\title{Generating bwa genome index}
\usage{
bwaIndex(group = c("sudo", "docker"), genome.folder = getwd(), ensembl.url,
  threads = 1)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{genome.folder, }{a character string indicating the folder where the indexed reference genome for bwa will be located}

\item{ensembl.url, }{a character string indicating the URL from ENSEMBL ftp for the unmasked genome sequence of interest}

\item{threads, }{a number indicating the number of cores to be used from the application}
}
\value{
The indexed bwa genome reference sequence
}
\description{
This function executes the docker container bwa1 where BWA is installed. The index is created using ENSEMBL genome fasta file. User needs to provide the URL for ENSEMBL genome located in the ENSEMBL ftp
}
\examples{
\dontrun{
    #running bwa index
    bwaIndex(group="sudo",genome.folder=getwd(),
    ensembl.url="ftp://ftp.ensembl.org/pub/release-87/fasta/homo_sapiens/dna/Homo_sapiens.GRCh38.dna.toplevel.fa.gz",
    threads=24)
}
}

