% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascImpute.R
\name{cascImpute}
\alias{cascImpute}
\title{Running CASC step1 identification of the optimal number of clusters}
\usage{
cascImpute(group = c("sudo", "docker"), data.folder, counts.matrix, drop.thre,
  cores)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{data.folder, }{a character string indicating the folder where tab separated file of cells counts is saved}

\item{counts.matrix, }{a character string indicating the  tab separated file of cells  counts}

\item{drop.thre, }{A number between 0 and 1, specifying the threshold to determine dropout values}

\item{cores, }{a integer specifying the number of cores used for parallel computation.}
}
\value{
...
}
\description{
This function executes CASC step1 identification of the optimal number of clusters
}
\examples{
\dontrun{
    #downloading fastq files
    system("wget http://130.192.119.59/public/singlecells_counts.txt.gz")
    system("gzip -d singlecells_counts.txt.gz")
    cascImpute(group="docker", data.folder=getwd(), counts.matrix="singlecells_counts.txt", drop.thre=0.5, cores=8)
}

}
