% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascKoptimization.R
\name{cascKoptimization}
\alias{cascKoptimization}
\title{Running CASC step1 identification of the optimal number of clusters}
\usage{
cascKoptimization(group = c("sudo", "docker"), scratch.folder,
  data.folder = getwd(), counts.matrix, permutations = 20,
  blocks.permutations = 2, core = 0, bootstrap.fraction = 10, k.min,
  k.max, totIdentity = 80, clusterIdentity = 80)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{data.folder, }{a character string indicating the folder where comma separated file of cells log10 counts is saved}

\item{counts.matrix, }{a character string indicating the the name of csv  file of cells log10 counts}

\item{permutations, }{a integer indicating how main permutations of SIMLR are going to be executed.}

\item{blocks.permutations, }{a integer indicating how many SIMLR permutation are going to be run in parallel. e.g. 10 indicates that 100 permutatin will be run in 10 groups of 10 permutations.}

\item{core, }{a integer between 0 to 1 to define the fraction of cores to be used by SIMLR, default 0, other values will generate an error in specific hadrware configurations.}

\item{bootstrap.fraction, }{a integer between 1 and 100, which indicate the fraction of cell to be removed during boostrap.}

\item{k.min, }{min number of clusters.}

\item{k.max, }{min number of clusters.}

\item{totIdentity, }{an integer between 1 to 100 which indicates the percentage of identity required to consider the results of two permutations identical.}

\item{clusterIdentity, }{an integer between 1 to 100 which indicates the percentage of identity required to consider two clusters identical.}
}
\value{
...
}
\description{
This function executes CASC step1 identification of the optimal number of clusters
}
\examples{
\dontrun{
    #downloading fastq files
    system("wget http://130.192.119.59/public/log10_singlecells_counts.csv.gz")
    system("gzip -d log10_singlecells_counts.csv.gz")
    #running cascKoptimization
    cascKoptimization(group="docker", scratch.folder="/Users/raffaelecalogero/Desktop/scratch", data.folder=getwd(),
    counts.matrix="log10_singlecells_counts.csv", permutations=20, blocks.permutations=2, core=0, bootstrap.fraction=10, k.min=2, k.max=4, totIdentity=80, clusterIdentity=80)
}

}
\author{
Raffaele Calogero, Luca Alessandri
}
