% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascOutputReformat.R
\name{cascOutputReformat}
\alias{cascOutputReformat}
\title{Organizing in a unique table for each K the output of cascKoptimization function}
\usage{
cascOutputReformat(data.folder)
}
\arguments{
\item{data.folder, }{a character string indicating the cascKoptimization output folder containing K folders, SilhouetteValue.csv, cell.stability.pdf and silhouette.pdf}
}
\value{
Returns in each K folder a tab delimited file called summary_table.txt, including the SilhouetteParameters (Silhouette score for each cell: extraScore, intraScore, ClusterBelong, Neighbor, SilhouetteValue), SIMLR.cluster, cell.stability.score, SIMLR component 1, SIMLR component 2.
}
\description{
This function converts organize in a table each output of cascKoptimization function
}
\examples{
\dontrun{
    #downloading fastq files
    system("wget http://130.192.119.59/public/example.zip")
    unzip("example.zip")
    setwd("./example")
    cascOutputReformat(data.folder=getwd())

}
}
