% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascStep4.R
\name{cascSelFeatures}
\alias{cascSelFeatures}
\title{Running CASC step4, feature selection analysis}
\usage{
cascSelFeatures(group = c("sudo", "docker"), scratch.folder,
  data.folder = getwd(), counts.matrix, k, dispersion.min = NULL,
  dispersion.max = NULL, specific.genes = c("n", "y"))
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{data.folder, }{a character string indicating the folder where the results of cascKoptimization: Data, Results, output}

\item{counts.matrix, }{a character string indicating the the name of csv  file of cells log10 counts}

\item{k, }{optimal number of clusters.}

\item{dispersion.min, }{min dispersion.}

\item{dispersion.max, }{max dispersion.}

\item{specific.genes, }{character with two option:  n, y}
}
\value{
....
}
\description{
This function executes CASC identification of the minimal subset of genes suitable to keep the cluster discrimination.
}
\details{
The first step of the analysis requires  that  both dispersion.min  and dispersion.max are set to NULL. This analysis will produce a plot describing the set of gene features specific for each  cluster that can be use  to refine the clustering. Subsequently giving  the range of  dispersion.min  and dispersion.max extrapolated from the the  first analysis will produce a SIMLR clustering describing  the effect of  the  feature selection.
}
\examples{
\dontrun{
    #downloading fastq files
    system("wget http://130.192.119.59/public/log10_singlecells_counts.csv.gz")
    system("gzip -d log10_singlecells_counts.csv.gz")
    # without parameter range
    cascFeature(group="docker", scratch.folder="/Users/raffaelecalogero/Desktop/scratch", data.folder=getwd(),
    counts.matrix="log10_singlecells_counts.csv",  k=3, dispersion.min=NULL, dispersion.max=NULL, specific.genes="n")
    # indicating a parameter range
    cascFeature(group="docker", scratch.folder="/Users/raffaelecalogero/Desktop/scratch", data.folder=getwd(),
    counts.matrix="log10_singlecells_counts.csv",  k=3, dispersion.min=2, dispersion.max=10, specific.genes="n")
}

}
\author{
Raffaele Calogero, Luca Alessandri
}
