% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascStep1.R
\name{cascStep1}
\alias{cascStep1}
\title{Running CASC step1 identification of the optimal number of clusters}
\usage{
cascStep1(group = c("sudo", "docker"), scratch.folder,
  data.folder = getwd(), counts.matrix, permutations = 100,
  blocks.permutations = 10, core = 0, bootstrap.fraction = 10, k.min,
  k.max)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{data.folder, }{a character string indicating the folder where comma separated file of cells log10 counts is saved}

\item{counts.matrix, }{a character string indicating the the name of csv  file of cells log10 counts}

\item{permutations, }{a integer indicating how main permutations of SIMLR are going to be executed.}

\item{blocks.permutations, }{a integer indicating how many SIMLR permutation are going to be run in parallel. e.g. 10 indicates that 100 permutatin will be run in 10 groups of 10 permutations.}

\item{core, }{a integer between 0 to 1 to define the fraction of cores to be used by SIMLR, default 0, other values will generate an error in specific hadrware configurations.}

\item{bootstrap.fraction, }{a integer between 1 and 100, which indicate the fraction of cell to be removed during boostrap.}

\item{k.min, }{min number of clusters.}

\item{k.max, }{min number of clusters.}
}
\value{
a folder with the name of the cell data set under analysis, e.g. log10_singlecells_counts. In this folder there are a set of folder named by the number of k clusters in use. A file named SilhouetteValue.csv for cluster evaluation and a pdf ViolinPlot.pdf which is the representation of the SilhouetteValue.csv. best clusters are represented by a SilhouetteValue distribution skewed versus 1 values.
}
\description{
This function executes CASC step1 identification of the optimal number of clusters
}
\examples{
\dontrun{
    #downloading fastq files
    system("wget http://130.192.119.59/public/log10_singlecells_counts.csv.gz")
    system("gzip -d log10_singlecells_counts.csv.gz")
    #running casc step1
    cascStep1(group="docker", scratch.folder="/Users/raffaelecalogero/Desktop/scratch", data.folder=getwd(),
    counts.matrix="log10_singlecells_counts.csv", permutations=20, blocks.permutations=2, core=0, bootstrap.fraction=10, k.min=2, k.max=4)
}

}
\author{
Raffaele Calogero, Luca Alessandri
}
