% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascStep2.R
\name{cascStep2}
\alias{cascStep2}
\title{Running CASC step2 identification of the optimal number of clusters}
\usage{
cascStep2(group = c("sudo", "docker"), scratch.folder,
  data.folder = getwd(), totIdentity = 80, clusterIdentity = 80, k.min,
  k.max, counts.matrix)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{data.folder, }{a character string indicating the folder where comma separated file of cells log10 counts is saved}

\item{totIdentity, }{an integer between 1 to 100 which indicates the percentage of identity required to consider the results of two permutations identical.}

\item{clusterIdentity, }{an integer between 1 to 100 which indicates the percentage of identity required to consider two clusters identical.}

\item{k.min, }{min number of clusters, should be the same of those passes to cascStep1.}

\item{k.max, }{min number of clusters, should be the same of those passes to cascStep1.}

\item{counts.matrix, }{a character string indicating the the name of csv  file of cells log10 counts}
}
\value{
a folder with .....
}
\description{
This function executes CASC step2 identification of the optimal number of clusters
}
\examples{
\dontrun{
    #downloading fastq files
    system("wget http://130.192.119.59/public/log10_singlecells_counts.csv.gz")
    system("gzip -d log10_singlecells_counts.csv.gz")
    #running casc step1
    cascStep2(group="docker", scratch.folder="/Users/raffaelecalogero/Desktop/scratch", data.folder=getwd(),
    totIdentity=80, clusterIdentity=80, k.min=2, k.max=4, counts.matrix="log10_singlecells_counts.csv")
}

}
\author{
Raffaele Calogero, Luca Alessandri
}
