% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascStep4.R
\name{cascStep4}
\alias{cascStep4}
\title{Running CASC step4, step1 of feature selection}
\usage{
cascStep4(group = c("sudo", "docker"), scratch.folder,
  data.folder = getwd(), counts.matrix, k, dispersion.min = NULL,
  dispersion.max = NULL, specific.genes = c("n", "y"))
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{data.folder, }{a character string indicating the folder where comma separated file of cells log10 counts is saved}

\item{counts.matrix, }{a character string indicating the the name of csv  file of cells log10 counts}

\item{k, }{optimal number of clusters.}

\item{dispersion.min, }{min dispersion.}

\item{dispersion.max, }{max dispersion.}

\item{specific.genes, }{character with two option:  n, y}
}
\value{
....
}
\description{
This function executes CASC step1 identification of the optimal number of clusters
}
\examples{
\dontrun{
    #downloading fastq files
    system("wget http://130.192.119.59/public/log10_singlecells_counts.csv.gz")
    system("gzip -d log10_singlecells_counts.csv.gz")
    # without parameter range
    cascStep4(group="docker", scratch.folder="/Users/raffaelecalogero/Desktop/scratch", data.folder=getwd(),
    counts.matrix="log10_singlecells_counts.csv",  k=3, dispersion.min=NULL, dispersion.max=NULL, specific.genes="n")
    # indicating a parameter range
    cascStep4(group="docker", scratch.folder="/Users/raffaelecalogero/Desktop/scratch", data.folder=getwd(),
    counts.matrix="log10_singlecells_counts.csv",  k=3, dispersion.min=2, dispersion.max=10, specific.genes="n")
}

}
