% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casc.checkCountDepth.R
\name{checkCountDepth}
\alias{checkCountDepth}
\title{Running SCnorm  checkCountDepth}
\usage{
checkCountDepth(group = c("sudo", "docker"), data.folder = getwd(),
  counts.matrix, conditions = NULL, outputName, nCores = 8)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{data.folder, }{a character string indicating the folder where comma separated file of cells log10 counts is saved}

\item{counts.matrix, }{a character string indicating the the name of tab delimited file  file of cells un-normalized expression counts}

\item{conditions, }{vector of condition labels, this should correspond to the columns of the un-normalized expression matrix. If not provided data is assumed to come from same condition/batch.}

\item{outputName, }{specify the path and/or name of output files.}

\item{nCores, }{number of cores to use, default is detectCores() - 1.}
}
\value{
pdf with the cells counts distributions
}
\description{
This function executes CASC step1 identification of the optimal number of clusters
}
\examples{
\dontrun{
    #downloading fastq files
    system("wget http://130.192.119.59/public/singlecells_counts.txt.gz")
    system("gzip -d singlecells_counts.txt.gz")
    checkCountDepth(group="docker", data.folder=getwd(),
    counts.matrix="singlecells_counts.txt", conditions=NULL,
    outputName="singlecells_counts", nCores=8)
}
}
