% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chipseq.R
\name{chipseq}
\alias{chipseq}
\title{Detection and annotation of TF binding sites and histone marks, based on the workflow used in Galli et al Mol Cell. 2015 Oct 15;60(2):328-37}
\usage{
chipseq(
  group = c("sudo", "docker"),
  bam.folder = getwd(),
  sample.bam,
  ctrl.bam,
  scratch.folder = "/data/scratch",
  genome = c("hg19", "hg38", "mm9", "mm10"),
  read.size,
  tool = c("macs", "sicer"),
  macs.min.mfold = 10,
  macs.max.mfold = 30,
  macs.pval = "1e-5",
  sicer.wsize = 200,
  sicer.gsize = c(200, 600),
  sicer.fdr = 0.1,
  tss.distance = 0,
  max.upstream.distance = 10000,
  remove.duplicates = c("Y", "N")
)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{bam.folder, }{a character string indicating where bam files are located}

\item{sample.bam, }{a character string indicating the chipseq file under analysis}

\item{ctrl.bam, }{a character string indicating the control file, e.g. unspecific IgG, input DNA, etc.}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{genome, }{a character string indicating the genome used as reference for data generation. Available options: hg19, hg38, mm9, mm10}

\item{read.size, }{an integer indicating the length of the sequenced reads}

\item{tool, }{a character string indicating the peaks calling algorith. Available options: macs and sicer. Macs, v 1.14, is used to call TF peaks, as instead sicer, v 1.1, is used to call histone mark peaks}

\item{macs.min.mfold, }{an integer indicating the minimum enrichment ratio against background}

\item{macs.max.mfold, }{an integer indicating the maximum enrichment ratio against background}

\item{macs.pval, }{a character string, indicationg the pvalue cutoff to be used to filter peaks with low statistical significance.The number must be provided in scientific notation as the default value shows}

\item{sicer.wsize, }{an integer indicating the windows size to be used by sicer}

\item{sicer.gsize, }{an integer indicating the gap size to be used by sicer. Suggested values: H3K4Me3=200; H3K27Me3=600}

\item{sicer.fdr, }{an integer indicating the pvalue cutoff to be used to filter peaks with low statistical significance}

\item{tss.distance, }{an integer indicating the distance of TSS with respect to gene start}

\item{max.upstream.distance, }{an integer indicating the maximum distance to associate a gene ID to a peak}

\item{remove.duplicates, }{a character string indicating if duplicated reads have to be removed. Available options: Y, to remove douplicates, N to keep duplicates}
}
\value{
three files: dedup_reads.bam, which is sorted and duplicates marked bam file, dedup_reads.bai, which is the index of the dedup_reads.bam, and dedup_reads.stats, which provides mapping statistics
}
\description{
This function executes the docker container chipseq.8 and requires as input two bam files, one for the chipseq of interest and the other for control, e.g. mock, which can be generated with bwa function.
}
\examples{
\dontrun{
    system("wget http://130.192.119.59/public/SRR1172111.bam")#TEAD
    system("wget http://130.192.119.59/public/SRR1172110.bam")#igg
    system("wget http://130.192.119.59/public/SRR1592211.bam")#H3K27ac
    #running chipseq for macs
    chipseq(group="sudo",bam.folder=getwd(), sample.bam="SRR1172111.bam", ctrl.bam="SRR1172110.bam",
    scratch.folder="/data/scratch", genome="hg19", read.size=50,
    tool="macs", macs.min.mfold=10, macs.max.mfold=30, macs.pval="1e-5",
    sicer.wsize=200, sicer.gsize=200, sicer.fdr=0.10, tss.distance=0, max.upstream.distance=10000,
    remove.duplicates="N")

    #running chipseq for sicer H3K4Me3
    chipseq(group="sudo",bam.folder=getwd(), sample.bam="SRR1592211.bam", ctrl.bam="SRR1172110.bam",
    scratch.folder="/data/scratch", genome="hg19", read.size=50,
    tool="sicer", sicer.wsize=200, sicer.gsize=200, sicer.fdr=0.10,
    tss.distance=0, max.upstream.distance=10000,remove.duplicates="N")
}
}
