% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circAnnotations.R
\name{circAnnotations}
\alias{circAnnotations}
\title{Annotation of a list of circRNAs}
\usage{
circAnnotations(group = c("sudo", "docker"), scratch.folder, ciri.file,
  annotation.sources = c("circbase", "tscd"),
  genome.version = c("hg18", "hg19", "hg38"))
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{ciri.file, }{a list of circRNAs derived from a CIRI2 prediction analysis}

\item{annotation.sources, }{a vector of character strings indicating the circRNA databases to analyse}

\item{genome.version, }{a character string indicating the reference genome assembly. The function currently work with the hg18, hg19 and hg38 human genome assemblies}
}
\value{
The annotations of a list of circRNAs from different databases
}
\description{
This function executes the docker container ciri2 in the annotation mode to overlap a list of circRNAs with the annotations from circBase and TSCD
}
\examples{
\dontrun{

# Retrieve the example data
    system("wget https://github.com/carlo-deintinis/circhunter/archive/master.zip")
    system("unzip master.zip")

# Run the circAnnotations function
 circAnnotations(group = "docker", scratch.folder="/data/scratch", ciri.file=paste(getwd(),"/circhunter-master/CircHunter/data/circRNA_CRC.bed", sep=""), genome.version="hg19")

circAnnotations()
}
}
\author{
Nicola Licheri and Giulio Ferrero
}
