% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circrnaAnnotations.R
\name{circrnaAnnotations}
\alias{circrnaAnnotations}
\title{Annotation of a list of circRNAs}
\usage{
circrnaAnnotations(
  group = c("sudo", "docker"),
  scratch.folder,
  ciri.file,
  annotation.sources = c("circbase", "tscd", "cscd", "exorbase", "circ2disease",
    "circfunbase"),
  assembly = c("hg18", "hg19", "hg38", "mm9", "mm10")
)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{ciri.file, }{a list of circRNAs derived from a circRNAs prediction analysis}

\item{annotation.sources, }{a vector of character strings indicating the circRNA databases to analyse. Compatible databases id: circbase, cscd, exorbase, circ2disease, circfunbase, tscd.}

\item{assembly, }{a character string indicating the reference genome assembly. The function currently work with the hg18, hg19, and hg38, mm9, and mm10 genome assemblies.}
}
\value{
The annotations of a list of circRNAs from different databases
}
\description{
This function executes the docker container ciri2 in the annotation mode to overlap a list of circRNAs with the annotations from circBase, CSCD, ExoRBase, Circ2Disease, CircFunBase, and TSCD
}
\examples{
\dontrun{

# Retrieve the example data
    system("wget https://github.com/carlo-deintinis/circhunter/archive/master.zip")
    system("unzip master.zip")

# Run the circAnnotations function
 circAnnotations(group = "docker", scratch.folder="/data/scratch", ciri.file=paste(getwd(),"/circhunter-master/CircHunter/data/circRNA_CRC.bed", sep=""), assembly="hg19")

circAnnotations()
}
}
\author{
Nicola Licheri and Giulio Ferrero
}
