% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circrnaOverlapResults.R
\name{circrnaOverlapResults}
\alias{circrnaOverlapResults}
\title{Running circRNA prediction overlap}
\usage{
circrnaOverlapResults(
  group = c("sudo", "docker"),
  scratch.folder,
  input.folder,
  output.folder,
  min_support
)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{input.folder, }{string indicating the path of the folder containing the samples to overlap}

\item{output.folder, }{string indicating the path of the output folder}

\item{min_support, }{an integer representing the minimum number of algorithms detecting the circRNA.}
}
\value{
The list of circRNAs detected in at least the user-defined number of algorithms
}
\description{
This function executes the docker container docker4circ to overlap circRNA predictions obtained from multiple tools
}
\examples{
\dontrun{
}
}
\author{
Nicola Licheri
}
