% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circrnaReformat.R
\name{circrnaReformat}
\alias{circrnaReformat}
\title{Running circRNA reformat script}
\usage{
circrnaReformat(
  group = c("sudo", "docker"),
  scratch.folder,
  input.folder,
  output.folder,
  tool = c("acfs", "circexplorer", "circexplorer2", "circrnafinder", "ciri", "ciri2",
    "dcc", "findcirc2", "knife", "starchip", "uroborus")
)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{input.folder, }{string indicating the path of the folder which contains the input files to be reformatted}

\item{output.folder, }{string indicating the path to the reformatted circRNA prediction files}

\item{tool, }{string indicating the tool used to create the input files. Supported tools are: acfs, circexplorer, circexplorer2, circrnafinder, ciri, ciri2, dcc, findcirc2, knife, starchip, uroborus}
}
\value{
A circRNA list reformatted in a tab-delimited file storing the circRNA genomic coordinates
}
\description{
This function executes the docker container docker4circ to reformat circRNA prediction file produced by a certain tool in a common format to further analysis.
}
\examples{
\dontrun{
}
}
\author{
Nicola Licheri
}
