% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciri2MergePredictions.R
\name{ciri2MergePredictions}
\alias{ciri2MergePredictions}
\title{Function to merge different circRNA lists from CIRI 2}
\usage{
ciri2MergePredictions(group = c("sudo", "docker"), scratch.folder,
  data.folder, groups.file, min_reads = 2, min_reps = 0,
  min_avg = 10)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{data.folder, }{a character string indicating the data folder where the CIRI 2 output files are located}
}
\value{
Two tab-delimited tables reporting the transcript- and gene-level classification of a list of circRNAs
}
\description{
This function executes the docker container ciri2merge by running the merge of different list of circRNAs predicted by CIRI2  following a sample data files provided by the user. The function executes also a filter based on the number of back-splicing reads computed in each experiment and across replicates of the same biological condition.
}
\author{
Nicola Licheri and Giulio Ferrero
}
