% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciri_as.R
\name{ciriAS}
\alias{ciriAS}
\title{Running CIRI_AS tool for circRNAs structure prediction}
\usage{
ciriAS(
  group = c("sudo", "docker"),
  scratch.folder,
  sam.file,
  ciri.file,
  genome.file,
  annotation.file = NA
)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{sam.file, }{a character string indicating the path to the RNA-Seq alignment SAM/BAM file from BWA}

\item{ciri.file, }{string indicating the path to the list of circRNAs}

\item{genome.file, }{a character string indicating the path to the Fasta file of the reference genomic sequence (it should be the same reference indexed for the BWA alignment)}

\item{annotation.file, }{a character string indicating the path to the GTF/GFF file reporting the reference gene annotations}
}
\value{
The function returns the list of alternative circRNAs internal structures
}
\description{
This function executes the docker container docker4circ where CIRI_AS is installed
}
\examples{
\dontrun{

    #Download the example data
    system("wget https://sourceforge.net/projects/ciri/files/CIRI-AS/test_data_CIRI_AS.zip/download")

    system("mv download test_data_CIRI_AS.zip")
    system("unzip test_data_CIRI_AS.zip")

 # Run the ciriAS function
ciriAS(group = "docker", scratch.folder="/data/scratch", sam.file=paste(getwd,"/test_data_CIRI_AS/test.sam",sep=""), ciri.file=paste(getwd,"/test_data_CIRI_AS/test.ciri", genome.file=paste(getwd,"/test_data_CIRI_AS/chr1.fa", sep=""), annotation.file = paste(getwd,"/test_data_CIRI_AS/chr1.gtf", sep="")
}

}
\author{
Nicola Licheri and Giulio Ferrero
}
