% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterReorg.R
\name{clusterReorg}
\alias{clusterReorg}
\title{Reorganize Cluster}
\usage{
clusterReorg(group = c("sudo", "docker"), scratch.folder, data.folder,
  matrixName, nCluster, A, B, format, separator, sp = 0.8, pValue = 0.9,
  clusterPermErr = 0.05)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{data.folder, }{a character string indicating the folder where input data are located and where output will be written}

\item{matrixName, }{counts table name. Matrix data file must be in data.folder. The file MUST contain RAW counts, without any modification, such as log transformation, normalizatio etc.}

\item{nCluster, }{number of Cluster used in Kmeans to generate the clusters that you want to merge}

\item{A, }{first Cluster that has to be merged}

\item{B, }{second Cluster that has to be merged}

\item{format, }{matrix count format, "csv", "txt"#' @param B, second Cluster that has to be merged}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{sp, }{minimun number of percentage of cells that has to be in common between two permutation to be the same cluster.}

\item{pValue, }{higher is the value more correct is the algorithm to recognize clusters along all the permutation.Default=0.9.}

\item{clusterPermErr, error}{that can be done by each permutation in cluster number depicting.Default = 0.05}
}
\value{
will change all the files generated from permAnalysis algorithm in a new folder matrixName_Cluster_merged/
}
\description{
This function executes a ubuntu docker that merge two clusters
}
\examples{
\dontrun{
clusterReorg("sudo","/home/lucastormreig/CASC2.0/2.1_clusterReorg/scratch/","/home/lucastormreig/CASC2.0/2.1_clusterReorg/Data/","TOTAL",3,1,3,"csv",",")# 
}
}
\author{
Luca Alessandri , alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
