% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cascTable.R
\name{counts2log}
\alias{counts2log}
\title{CASC requires log10 counts table in comma delimited file}
\usage{
counts2log(counts.matrix, data.folder = getwd(), log.base = c(2, 10),
  type = c("txt", "csv"))
}
\arguments{
\item{counts.matrix, }{a character string indicating the the name of txt or csv  file of cells counts}

\item{data.folder, }{a character string indicating the folder where comma separated file of cells log10 counts will be saved}

\item{log.base, }{a integer value indicating the base fro the log transformation}

\item{type, }{a character value indicating the content of the file separator: csv comma, txt tab}
}
\value{
Returns a log10 or log2 comma separated file.
}
\description{
This function converts raw counts table in CASC suitable table.
}
\examples{
\dontrun{
    #downloading fastq files
    system("wget http://130.192.119.59/public/singlecells_counts.txt.gz")
    system("gzip -d singlecells_counts.txt.gz")
    counts2log(counts.matrix="singlecells_counts.txt", data.folder=getwd(), log.base=10, type="txt")

}
}
