% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deeptoolBwig.R
\name{deeptoolBwig}
\alias{deeptoolBwig}
\title{Create a bigWig from the output of bowtie2}
\usage{
deeptoolBwig(group = c("sudo", "docker"), fastq.folder = getwd(),
  scratch.folder = "/data/scratch")
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{fastq.folder, }{a character string indicating where bam, bai files  are located}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}
}
\value{
sorted.bw
}
\description{
This function executes the docker container bowtie2 embedding deeptools the output of bowtie2 sorted.bam is used to generate the bigWig
}
\examples{
\dontrun{
    #downloading fastq files
system("wget http://130.192.119.59/public/test_R1.fastq.gz")
system("wget http://130.192.119.59/public/test_R2.fastq.gz")
library(docker4seq)
#running bowtie nostrand pe
deeptoolBwig(group="docker",fastq.folder=getwd(), scratch.folder="/data/scratch/")

}
}
\author{
Raffaele Calogero
}
