% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDocker.R
\name{dockerTest}
\alias{dockerTest}
\title{Testing if Docker is installed}
\usage{
dockerTest()
}
\value{
a character string indicating the version of the docker installed in the system
}
\description{
This function check that docker is installed
}
\examples{
 dockerTest()
}
