% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadContainers.R
\name{downloadContainers}
\alias{downloadContainers}
\title{Download for the first time all containers embedded in the workflows}
\usage{
downloadContainers(group = "docker", containers.file = "all")
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{containers.file, }{a character string with the name of the file which indicate which are the initial set of containers to be downloaded. Initally the set is given by a file located in the folder containers of docker4seq package: "all", "rnaseq", "ncrnaseq", "chipseq"}
}
\description{
This is a functin that preapre the docker environment to be used for the first time the docker4seq is installed.
}
\examples{
\dontrun{
    #running runDocker
     downloadContainers(group="docker", containers.file)

}
}
\author{
Raffaele Calogero
}
