% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentPower.R
\name{experimentPower}
\alias{experimentPower}
\title{A wrapper function for experiment_power from RnaSeqSampleSize Bioconductor package}
\usage{
experimentPower(
  group = c("sudo", "docker"),
  filename,
  replicatesXgroup = 3,
  FDR = 0.1,
  genes4dispersion = 200,
  log2fold.change = 1,
  output.folder = getwd()
)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{filename, }{a character string indicating the name of the count table file}

\item{replicatesXgroup, }{an integer indicating the number of samples used in each group}

\item{FDR, }{false discovery rate}

\item{genes4dispersion, }{an integer indicating the number of genes used in estimation of read counts and dispersion distribution}

\item{log2fold.change, }{an integer indicating the minimum log2 fold change for prognostic genes between two groups}

\item{output.folder, }{a string indicating the path where to save the output file}
}
\value{
a string with the requested informations. The string is also saved in a file: power_evaluation.txt
}
\description{
This function evaluate the statistical power of a pilot experiment
}
\examples{
\dontrun{
 system("wget 130.192.119.59/public/test.analysis.zip")
 unzip("test.analysis.zip")
 setwd("test.analysis")
 library(docker4seq)
 experimentPower(group="docker",filename="_counts.txt",replicatesXgroup=7,
 FDR=0.1, genes4dispersion=200, log2fold.change=1)
}
}
\author{
Raffaele Calogero
}
