% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastq2fasta.R
\name{fastq2fasta}
\alias{fastq2fasta}
\title{A function to handle the conversion from fastq to fasta}
\usage{
fastq2fasta(group = c("sudo", "docker"), scratch.folder, data.folder)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{data.folder, }{a character string indicating the folder where input data are located and where output will be written}
}
\value{
Returns sample.fasta file in the data.folder
}
\description{
This function executes a ubuntu docker that convert a fastq.gz file in a fasta file using seqtk
}
\examples{
\dontrun{
    #running fastq2fasta
    fastq2fasta(group="docker", scratch.folder="/data/scratch", data.folder=getwd())
}

}
\author{
Raffaele A Calogero, raffaele.calogero [at] unito [dot] it, University of Torino. Italy
}
