% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterCounts.R
\name{filterCounts}
\alias{filterCounts}
\title{Filter a count table using a table of DE from wrapperDeseq2}
\usage{
filterCounts(data.folder, type = c("gene", "isoform", "mirna"))
}
\arguments{
\item{data.folder, }{a character string indicating the paths of rnaseqCounts/mirnaCounts output folders}

\item{type, }{character with three options: gene, isoform, mirna.}
}
\value{
Returns counts, fpkm, tpm data frames for gene and isoforms in countsDE.txt, log2fpkmDE.txt and in log2TPMDE.txt
}
\description{
This function generates counts, FPKM and TPM tables including only the differentially expressed genes present in the set of DE generated with wrapperDeseq2.
}
\examples{
\dontrun{
    system("wget 130.192.119.59/public/test.analysis.zip")
    unzip("test.analysis.zip")
    setwd("test.analysis")
    library(docker4seq)
    wrapperDeseq2(output.folder=getwd(), group="docker", experiment.table="_counts.txt", log2fc=1,
    fdr=0.1, ref.covar="Cov.1", type="gene", batch=FALSE))

    filterCounts(data.folder=getwd(), type="gene")

}
}
\author{
Raffaele Calogero
}
