% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatkDna.R
\name{gatkDNA}
\alias{gatkDNA}
\title{Running realignment and recalibration, GATK}
\usage{
gatkDNA(group = c("sudo", "docker"), bam.folder = getwd(),
  scratch.folder = "/data/scratch", genome.folder, gatk.path, dbsnp.file,
  g1000.file, threads = 1)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{bam.folder, }{a character string indicating where bam files generated with bwa.R are located}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{genome.folder, }{a character string indicating the folder where the indexed reference genome for bwa is located}

\item{gatk.path, }{a character string indicating the path to GenomeAnalysisTK-X.X-0.tar.bz2}

\item{dbsnp.file, }{a character string indicating the name of dbSNP vcf located in the genome folder. The dbSNP vcf, dbsnp_138.b37.vcf.gz, can be downloaded from ftp://gsapubftp-anonymous@ftp.broadinstitute.org/bundle/b37}

\item{g1000.file, }{a character string indicating the name of 1000 genome vcf located in the genome folder. The 1000 genomes vcf, Mills_and_1000G_gold_standard.indels.b37.vcf.gz, can be downloaded from ftp://gsapubftp-anonymous@ftp.broadinstitute.org/bundle/b37/}

\item{threads, }{a number indicating the number of cores to be used from the application}
}
\value{
three files: dedup_reads.bam, which is sorted and duplicates marked bam file, dedup_reads.bai, which is the index of the dedup_reads.bam, and dedup_reads.stats, which provides mapping statistics
}
\description{
This function executes the docker container snv.1 where GATK software is used to do INDEL realignment and quality recalibration. This analysis is required only to run mutect1
}
\examples{
\dontrun{
    #downloading fastq files
    system("wget http://130.192.119.59/public/test_R1.fastq.gz")
    system("wget http://130.192.119.59/public/test_R2.fastq.gz")
    #running bwa
    gatkDna(group="sudo",bam.folder=getwd(), scratch.folder="/data/scratch",
    genome.folder="/data/scratch/hg19_exome", gatk.path=getwd(), dbsnp.file="dbsnp_138.b37.vcf.gz",
    g1000.file="Mills_and_1000G_gold_standard.indels.b37.vcf.gz",
    threads=24, sample.id="exome")
}
}

