% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGeneInfo.R
\name{getInfo}
\alias{getInfo}
\title{Downloading the full set of genes information from HGNC}
\usage{
getInfo(
  group = c("sudo", "docker"),
  data.folder = getwd(),
  url,
  type = c("pubmed", "omim")
)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{data.folder, }{a character string indicating where downloaded data will be stored}

\item{url, }{the url to download the full dataset}

\item{type, }{PUBMED or OMIM}
}
\value{
one file files: gene_with_protein_product.txt
}
\description{
This function retrieves the full table of curated genes annotations
}
\examples{
\dontrun{
    getInfo(group="docker",data.folder=getwd(),
    url="ftp.ebi.ac.uk/pub/databases/genenames/new/tsv/locus_types/gene_with_protein_product.txt", type="pubmed")

}
}
\author{
Raffaele Calogero
}
