% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashclone.R
\name{hashclone}
\alias{hashclone}
\title{HashClone running A function to handle a hashclone docker container}
\usage{
hashclone(group = c("sudo", "docker"), scratch.folder,
  data.folder = getwd(), kmer, hash, coll, threashold = 0.1,
  spike = "null", input.files)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{data.folder, }{a character string indicating the folder where input data are located and where output will be written}
}
\description{
This function executes  HashClone algorithm developed to identify the set of clonality markers during the patient follow-up in order to quantify the minimal residual disease.
}
\examples{
\dontrun{
    #running skeleton
    skeleton(group="docker", scratch.folder="/Users/raffaelecalogero/Desktop/scratch", 
    data.folder=getwd())
}

}
\author{
Beccuti Marco, Greta Romano, Francesca Cordero, Raffaele Calogero, beccuti@di.unito.it, Computer Science Department Univ. of Turin.
}
