% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapBase.R
\name{heatmapBase}
\alias{heatmapBase}
\title{hfc, heatmap for clustering}
\usage{
heatmapBase(group = c("sudo", "docker"), scratch.folder, data.folder,
  matrixName, format, separator, geneNameControl = 1, status)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{data.folder, }{a character string indicating the folder where input data are located and where output will be written}

\item{matrixName, }{counts table name. Matrix data file must be in data.folder. The file MUST contain RAW counts, without any modification, such as log transformation, normalizatio etc.}

\item{format, }{matrix count format, "csv", "txt"}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{geneNameControl, }{0 if the matrix has gene name without ENSEMBL code.1 if the gene names is formatted like this : ENSMUSG00000000001:Gnai3. If the gene names is only ensamble name you have to run SCannoByGtf before start using this script.}

\item{status, }{0 if is raw count, 1 otherwise}
}
\value{
A heatmap.
}
\description{
This function generate heatmap and other plot based on clustering and on a specific gene list
}
\examples{
\dontrun{
system("wget http://130.192.119.59/public/heatmap_test.zip")
system("unzip heatmap_test.zip")
setwd("heatmap_test")
heatmapBase(group="docker",scratch.folder="/data/scratch",data.folder=getwd(),matrixName="DEfiltered__log2TPM",format="txt",separator="\\t",geneNameControl=1,status=hcl1)
}
}
\author{
Luca Alessandri , alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
