% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaply.R
\name{heatmaply}
\alias{heatmaply}
\title{Function to produce an interactive heatmap using plotly}
\usage{
heatmaply(group = c("docker", "sudo"), scratch.folder, count.table,
  gene.list, output.folder, separator = c("TAB", "COMMA", "SPACE"),
  status = c("raw", "log"), lower.range = NA, upper.range = NA)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{count.table, }{a character string indicating the path of the counts table file}

\item{gene.list, }{a character string indicating the path of the file containing the genes to include in the heatmap}

\item{output.folder, }{a character string indicating the path of the output folder}

\item{separator, }{a character string indicating the separator character in the count table. Allowed characters are TAB, COMMA and SPACE}

\item{status, }{a character string, 'raw' if the data are raw counts, 'log' otherwise}
}
\description{
This function generates an heatmap using a count table and a specific gene list
}
\examples{
\dontrun{

}
}
\author{
Nicola Licheri
}
