% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indropCounts.R
\name{indropCounts}
\alias{indropCounts}
\title{A function to handle a indrop V2 single cell data}
\usage{
indropCounts(group = c("sudo", "docker"), scratch.folder, fastq.folder,
  index.folder)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{fastq.folder, }{a character string indicating the folder where input data are located and where output will be written}

\item{index.folder, }{a character string indicating the folder where transcriptome index was created with salmonIndex.}
}
\description{
This function executes a docker that produces as output the sinngle cell counts from V2 indrop single cell sequencing
}
\examples{
\dontrun{
library(docker4seq)
#running salmonCounts
}

}
\author{
Raffaele Calogero and Riccardo Panero, raffaele.calogero [at] unito [dot] it, Bioinformatics and Genomics unit, University of Torino Italy
}
