% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kraken.R
\name{kraken}
\alias{kraken}
\title{A function executing kraken}
\usage{
kraken(
  group = "docker",
  scratch.folder,
  fastq.folder,
  genome.folder,
  fastq.prefix,
  threads
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{fastq.folder, }{a character string indicating the folder where input fastq(s) are located and where output will be written}

\item{genome.folder, }{a character string indicating the folder where the indexed kraken db is located}

\item{fastq.prefix, }{a character string indicating the PREFIX of the fastq file(s): e.g. prova_R1.fastq.gz fastq.prefix="prova"}

\item{threads, }{a number indicating the number of cores to be used from the application}
}
\description{
This function executes a ubuntu docker that embed kraken2
}
\examples{
\dontrun{
    #running skeleton
    kraken(group="docker", scratch.folder="/data/scratch", 
    fastq.folder=getwd(), genome.folder="/data/genomes/minikraken",
    fastq.prefix="unmapped", threads=8)
}

}
\author{
Raffaele A Calogero, raffaele.calogero [at] unito [dot] it, University of Torino
}
