% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masc2.R
\name{macs2}
\alias{macs2}
\title{A function to handle a MACS2 containier}
\usage{
macs2(group = c("sudo", "docker"), control.bam, chipseq.bam,
  experiment.name)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{control.bam, }{aa character string indicating the path to the control bam file. IMPORTANT control.bam and chipseq.bam are in the same folder}

\item{chipseq.bam, }{a character string indicating the path to the chipseq bam file. IMPORTANT control.bam and chipseq.bam are in the same folder}

\item{experiment.name, }{a character string indicating the prefix for MCS2 output}
}
\description{
This function executes a MACS2 docker that produces as output peaks call
}
\examples{
\dontrun{
    #running MACS
    macs2(group="docker", control.bam=paste(getwd(),"igg_dedup_reads.bam", sep="/"), 
           chipseq.bam=paste(getwd(),"prdm51_dedup_reads.bam", sep="/"), experiment.name="prdm51_igg")
}

}
\author{
raffaele.calogero [at] unito [dot] it, University of Torino
}
