% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mergeData.R
\name{mergeData}
\alias{mergeData}
\title{Function to merge different circRNA lists from CIRI 2}
\usage{
mergeData(
  group = c("sudo", "docker"),
  scratch.folder,
  data.folder,
  samples.ids,
  covariates,
  covariate.order,
  extension,
  column_index
)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{data.folder, }{a character string indicating the data folder where the file to merge are located}

\item{samples.ids, }{a character vector indicating the identifiers of the samples}

\item{covariates, }{a character vector indicating the classes of the samples}

\item{covariate.order, }{a character vector indicating a vector reporting the covariate classes ordered as desidered in the output file}

\item{extension, }{a character string indicating the filename extension of the files that have to merge}

\item{column_index, }{an integer value > 1 indicating which column values have to been reported in the output file}
}
\value{
Two tab-delimited tables reporting the BS supporting reads and the coordinates of the filtered circRNAs are reported
}
\description{
This function executes the docker container ciri2merge by running the merge of different lists of circRNAs predicted by CIRI2  following a sample data files provided by the user. The function executes also a filter based on the number of back-splicing reads computed in each experiment and across replicates of the same biological condition.
}
\examples{
\dontrun{

}
}
\author{
Nicola Licheri and Giulio Ferrero
}
