% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mirnaCounts.R
\name{mirnaCounts}
\alias{mirnaCounts}
\title{Counting miRNAs, Cordero et al. PLoS One. 2012;7:e31630}
\usage{
mirnaCounts(
  group = c("sudo", "docker"),
  fastq.folder = getwd(),
  scratch.folder = "/data/scratch",
  mirbase.id = c("hsa", "mmu"),
  download.status = FALSE,
  adapter.type = c("ILLUMINA", "NEB"),
  trimmed.fastq = FALSE
)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{fastq.folder, }{a character string indicating where mirna fastq are located}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{mirbase.id, }{a character string indicating the mirbase prefix for the organism of interest, e.g. hsa for human or mmu for mouse}

\item{download.status, }{a boolean logical variable indicating if the latest mirbase database will be downloaded or the local mirbase 21 will be used. Default is FALSE}

\item{adapter.type, }{a character string. Two options: \code{"ILLUMINA"} or \code{"NEB"}, depending to which miRNA library prep was used: ILLUMINA or NEB}

\item{trimmed.fastq, }{a boolean logical variable indicating if trimmed fastq are saved. Default is FALSE}
}
\value{
one file: annotated_genes.results, which is the annotated version of gene.results.
}
\description{
This function executes the docker container mrna8, which allows miRNAs counting.
}
\examples{
\dontrun{
   system("wget 130.192.119.59/public/test.mirnaCounts.zip")
   unzip("test.mirnaCounts.zip")
   setwd("test.mirnaCounts")
   library(docker4seq)
   mirnaCounts(group="docker",fastq.folder=getwd(),
             scratch.folder="/data/scratch",
             mirbase.id="hsa",download.status=FALSE,
             adapter.type="NEB", trimmed.fastq=FALSE)

}
}
\author{
Raffaele Calogero
}
