% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncosnpIllumina.R
\name{oncosnp}
\alias{oncosnp}
\title{A function to handle a docker containier executing CNV calculation. Only supporting hg19}
\usage{
oncosnp(
  group = c("sudo", "docker"),
  data.folder,
  scratch.folder,
  sample.name = NULL,
  blood.name = NULL
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{data.folder, }{a character string indicating the folder where input data are located and where output will be written}

\item{scratch.folder, }{a character string indicating the scratch folder for temporary operations}

\item{sample.name, }{a character string indicating snp data file name}

\item{blood.name, }{a character string indicating snp data file name for blood snp}
}
\value{
cnv1.cnvs, cnv2,cnvs and other oncosnp elements, see oncosnp help page for further information
}
\description{
This function executes a ubuntu docker that produces as output oncosnp and request as input a tabl delimited file with Name	Chromosome	Position	Log R Ratio	B Allele Freq, where Name is the snp id from Illumina arrays, Chromosome is the chr, POsition is the SNP location in the chr Log R Ratio	B Allele Freq are genrated from genomestudio
}
\examples{
\dontrun{
    system("wget http://130.192.119.59/public/testcnv.zip")
    #running fastqc
    oncosnp(group="docker", data.folder=getwd(), scratch.folder="/data/scratch", sample.name="sampleXX.txt", blood.name=NULL)
}

}
\author{
Raffaele Calogero, raffaele.calogero [at] unito [dot] it, University of Torino
}
