% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oncosnpAnnotation.R
\name{oncosnpAnnotation}
\alias{oncosnpAnnotation}
\title{A function to associate CNV generated by oncSNP starting from log R ratio and B allel frequences, generated by SNP arrays, to gene symbols}
\usage{
oncosnpAnnotation(group = c("sudo", "docker"), data.folder, genome.folder)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{data.folder, }{a character string indicating the folder where input data (.cnvs) generated with oncoSNP are located and where output will be written}

\item{genome.folder, }{a character string indicating the path of the folder where ENSEMBL genome.gtf is located, IMPORTANT use the same genome assembly used to design the array.}
}
\value{
a file for each .cnvs, e.g. annotation_XXXX.cnvs.txt
}
\description{
This function executes a ubuntu docker that uses as input the cnvs files produced by oncoSNP and add CN and LOH (check oncoSNP readme for more info) fields (rank 5 only) to the genes characterized by having genomics cohordinates included in the CNV region
}
\examples{
\dontrun{
    #running oncosnpAnnotation
system("wget http://130.192.119.59/public/test_oncosnp.zip")
system("unzip test_oncosnp.zip")
oncosnpAnnotation(group="docker", data.folder="./test_oncosnp/oncosnp_out", genome.folder="./test_oncosnp/hg19")
}

}
\author{
Raffaele Calogero, raffaele.calogero[at]unito[dot]it, University of Torino, Italy
}
