% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\title{generating a PCA from counts, FPKM and TPM tables from rnaseqCounts outuputs}
\usage{
pca(experiment.table = "_counts.txt", type = c("counts", "FPKM", "TPM"),
  covariatesInNames = FALSE, principal.components = c(1, 2),
  legend.position = c("bottom", "bottomleft", "left", "topleft", "top",
  "topright", "right", "center"), pdf = TRUE, output.folder = getwd())
}
\arguments{
\item{experiment.table, }{a character string indicating the counts, FPKM or TPM table file name}

\item{type, }{a character value indicating the content of the file: counts, FPKM or TPM}

\item{covariatesInNames, }{a boolean value indicating if covariates are inserted after \_ in the filename}

\item{principal.components, }{a numerical vector with two values indicating the principal components to be plotted}

\item{legend.position, }{a character string indicating the location of the covariates legend}

\item{pdf, }{a boolean value indicating if results has to be saved in a pdf}

\item{output.folder, }{output foldeer}
}
\value{
Returns a PCA plot
}
\description{
This function generates PCA plot from counts, FPKM and TPM tables from rnaseqCounts outuputs.
}
\examples{
\dontrun{
    # Example of the header of the counts, FPKM or TPM contaning covariate
    # no covariate in the name ctrl1 ctrl2 trt1 trt2
    # no covariate in the name ctrl1_c ctrl2_c trt1_t trt2_t
    pca(experiment.table="_log2FPKM.txt", type="FPKM", legend.position="topleft",
    covariatesInNames=FALSE, principal.components=c(1,2),
    pdf = TRUE, output.folder=getwd())

}
}
