% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permAnalysis.R
\name{permAnalysis}
\alias{permAnalysis}
\title{A skeleton function to handle a docker containier}
\usage{
permAnalysis(group = c("sudo", "docker"), scratch.folder, data.folder,
  matrixName, range1, range2, format, separator, sp)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{data.folder, }{a character string indicating the folder where input data are located and where output will be written}

\item{matrixName, }{counts table name. Matrix data file must be in data.folder. The file MUST contain RAW counts, without any modification, such as log transformation, normalizatio etc.}

\item{range1, }{DA AGGIUNGERE}

\item{range2, }{DA AGGIUNGERE}

\item{format, }{DA AGGIUNGERE}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{sp, }{DA AGGIUNGERE}
}
\value{
output will be in the same format and with the same separator of input.
}
\description{
This function executes a ubuntu docker that produces as output helloworld.txt file
}
\examples{
\dontrun{
permAnalysis("sudo","path/to/scratch","path/to/data","TOTAL",3,4,"csv",",",0.8)# 
}
}
\author{
Luca Alessandri , alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
