% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationClustering.R
\name{permutationClustering}
\alias{permutationClustering}
\title{A skeleton function to handle a docker containier}
\usage{
permutationClustering(group = c("sudo", "docker"), scratch.folder,
  data.folder, matrixName, nPerm, permAtTime, percent, range1, range2, format,
  separator, logTen, clustering)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{data.folder, }{a character string indicating the folder where input data are located and where output will be written}

\item{matrixName, }{counts table name. Matrix data file must be in data.folder. The file MUST contain RAW counts, without any modification, such as log transformation, normalizatio etc.}

\item{nPerm, }{DA AGGIUNGERE}

\item{permAtTime, }{DA AGGIUNGERE}

\item{percent, }{DA AGGIUNGERE}

\item{range1, }{DA AGGIUNGERE}

\item{range2, }{DA AGGIUNGERE}

\item{format, }{DA AGGIUNGERE}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{logTen, }{DA AGGIUNGERE}

\item{clustering, }{DA AGGIUNGERE}
}
\value{
output will be in the same format and with the same separator of input.
}
\description{
This function executes a ubuntu docker that produces as output helloworld.txt file
}
\examples{
\dontrun{
 permutationClustering("sudo","/home/lucastormreig/CASC2.0/permutationClustering/scratch/","/home/lucastormreig/CASC2.0/permutationClustering/Data/","TOTAL",4,2,10,3,4,"csv",",",0,"SIMLR")# 
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
