% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandsReformat.R
\name{prepare4expands}
\alias{prepare4expands}
\title{A function to handle a docker containier preparing the file needed for expanse}
\usage{
prepare4expands(group = c("sudo", "docker"), data.folder)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{data.folder, }{a character string indicating the folder where input data are located and where output will be written}
}
\value{
cnv.txt and snv.txt
}
\description{
This function executes a ubuntu docker that produces the cnv.txt and snv.txt needed for expanse
}
\examples{
\dontrun{
    system("wget http://130.192.119.59/public/prepare4expanse.zip")
    system("unzip prepare4expanse.zip")
    prepare4expands(group="docker", data.folder=paste(getwd(), "prepare4expanse", sep="/"))
}

}
\author{
Raffaele Calogero, raffaele.calogero [at] unito [dot] it, University of Torino
}
