% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubmedNet.R
\name{pubmedNet}
\alias{pubmedNet}
\title{Creating networks on the basis of the links between genes ids and pubmedid}
\usage{
pubmedNet(group = c("sudo", "docker"), file)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{file, }{a character string indicating the output of DESeq or ANOVAlike to be used. In the same folder must be present also genes.edges file.}
}
\value{
one graph plot
}
\description{
This function build networks using the genes.edges generated by getInfo function and create a graph using the genes present in a file generated by DESeq or ANOVAlike
}
\examples{
\dontrun{
    pubmedNet(group="docker", file=paste(getwd(),"DE_anova_rankedBySD_filteredByPositivePatient.txt", sep="/"))

}
}
\author{
Raffaele Calogero
}
