% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqCounts.R
\name{rnaseqCounts}
\alias{rnaseqCounts}
\title{Running RNAseq counting workflow for a single sample}
\usage{
rnaseqCounts(group = "sudo", fastq.folder = getwd(),
  scratch.folder = "/data/scratch", threads = 4,
  adapter5 = "AATGATACGGCGACCACCGAGATCTACACTCTTTCCCTACACGACGCTCTTCCGATCT",
  adapter3 = "AATGATACGGCGACCACCGAGATCTACACTCTTTCCCTACACGACGCTCTTCCGATCT",
  seq.type = "pe", min.length = 40,
  genome.folder = "/data/scratch/hg38star", strandness = "none",
  org = "hg38", truncating.expected.counts = FALSE,
  annotation.type = "rsemENSEMBL")
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{fastq.folder, }{a character string indicating where gzip fastq files are located}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{threads, }{a number indicating the number of cores to be used from the application}

\item{adapter5, }{a character string indicating the fwd adapter}

\item{adapter3, }{a character string indicating the rev adapter}

\item{seq.type, }{a character string indicating the type of reads to be generated by the sequencer. Two options: \code{"se"} or \code{"pe"} respectively for single end and pair end sequencing.}

\item{min.length, }{a number indicating minimal length required to return a trimmed read
#params rsemstar}

\item{genome.folder, }{a character string indicating the folder where the indexed reference genome for bwa is located. IMPORTANT the present function only suport genomic indexes made using ensembl genom and the corresponding gtf}

\item{strandness, }{a character string indicating the type ofsequencing protocol used for the analysis. Three options: \code{"none"}, \code{"forward"}, \code{"reverse"} respectively for non strand selection, forward for Illumina strandness protocols, reverse for ACCESS Illumina protocol
#params rsemanno}

\item{org, }{a character string indicating the genome assembly used for mapping and counting with \code{"rsemstar"} function}

\item{truncating.expected.counts, }{a boolean logical variable indicating if the expected counts calculated by RSEM need to be converted in integer to be compliant with differnetial expression Bioconductor packages as DESeq2. Default is FALSE}

\item{annotation.type, }{a character string. Two options: \code{"rsemUCSC"} or \code{"rsemENSEMBL"}, depending to which database was used in rsemstar for counting: UCSC or ENSEMBL}
}
\value{
Returns the output of skewer, rsemstar, rsemanno
}
\description{
This function executes a set of docker containers allowing the generation of gene and isoforms counts for a single sample.
#params skewer
}
\examples{
\dontrun{
    system("wget http://130.192.119.59/public/test_R1.fastq.gz")
    system("wget http://130.192.119.59/public/test_R2.fastq.gz")
    rnaseqCounts(group="sudo",fastq.folder=getwd(), scratch.folder="/data/scratch",
    adapter5="AATGATACGGCGACCACCGAGATCTACACTCTTTCCCTACACGACGCTCTTCCGATCT",
    adapter3="AATGATACGGCGACCACCGAGATCTACACTCTTTCCCTACACGACGCTCTTCCGATCT",
    seq.type="pe", threads=10,  min.length=40,
    genome.folder="/data/scratch/hg38star", strandness="none",
    org="hg38", truncating.expected.counts=FALSE, annotation.type="rsemENSEMBL")
}
}

