% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsemAnnotate.R
\name{rsemanno}
\alias{rsemanno}
\title{Annotating RSEM gene.results using ENSEMBL annotation}
\usage{
rsemanno(
  group = c("sudo", "docker"),
  rsem.folder = getwd(),
  scratch.folder = "/data/scratch",
  org = c("hg19", "hg38", "mm10", "mm9"),
  truncating.expected.counts = FALSE,
  protein.anno = FALSE
)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{rsem.folder, }{a character string indicating where gene.results and isoforms.results are located}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{org, }{a character string indicating the genome assembly used for mapping and counting with \code{"rsemstar"} function}

\item{truncating.expected.counts, }{a boolean logical variable indicating if the expected counts calculated by RSEM need to be converted in integer to be compliant with differnetial expression Bioconductor packages as DESeq2. Default is FALSE}

\item{protein.anno, }{a boolean logical variable indicating if instead of gene SYMBOL SWISSPROT symbol are used. This option is useful for integrating transcriptomics data with proteomics data}
}
\value{
one file: annotated_genes.results, which is the annotated version of gene.results.
}
\description{
This function executes the docker container annotate.1, where Bioconductor is used to annotated gene.results output of RSEM using ENSEMBL annotation
}
\examples{
\dontrun{
    #downloading fastq files
    system("wget http://130.192.119.59/public/genes.results.gz")
    gzip -d genes.results.gz
    #running rsemanno
    rsemanno(group="docker",rsem.folder=getwd(), scratch.folder="/data/scratch",
    org="hg38", truncating.expected.counts=FALSE,
    protein.anno=FALSE)

}
}
