% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsemBw.R
\name{rsemBw}
\alias{rsemBw}
\title{Creating a bigwig using RSEM}
\usage{
rsemBw(
  group = c("sudo", "docker"),
  bam.folder = getwd(),
  scratch.folder = "/data/scratch"
)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{bam.folder, }{a character string indicating where BAM SORTED file is located}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}
}
\value{
output.bw, which is the bigwig
}
\description{
This function executes the docker container rsemstar where RSEM is installed and create a bigwig file for genomic data visualization
}
\examples{
\dontrun{
    #downloading fastq files
    rsemBw(group="docker",bam.folder=getwd(), scratch.folder="/data/scratch")


}
}
