% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsemstarIndex.R
\name{rsemstarIndex}
\alias{rsemstarIndex}
\title{Generating rsem-star genome index}
\usage{
rsemstarIndex(
  group = c("sudo", "docker"),
  genome.folder = getwd(),
  ensembl.urlgenome = NULL,
  ensembl.urlgtf = NULL,
  threads = 1
)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{genome.folder, }{a character string indicating the folder where the indexed reference genome for STAR will be located}

\item{ensembl.urlgenome, }{a character string indicating the URL from ENSEMBL ftp for the unmasked genome sequence of interest}

\item{ensembl.urlgtf, }{a character string indicating the URL from ENSEMBL ftp for the GTF for genome of interest}

\item{threads, }{a number indicating the number of cores to be used from the application}
}
\value{
The index of the reference genomic sequence for STAR analysis
}
\description{
This function executes the docker container rsem-star1 where RSEM and STAR are installed. The index is created using ENSEMBL genome fasta file. User needs to provide the URL for ENSEMBL genome located in the ENSEMBL ftp
}
\examples{
\dontrun{
    #running rsemstar index for human
    rsemstarIndex(group="sudo",genome.folder="/data/scratch/hg38star",
    ensembl.urlgenome=
    "ftp://ftp.ensembl.org/pub/release-87/fasta/homo_sapiens/dna/Homo_sapiens.GRCh38.dna.toplevel.fa.gz",
    ensembl.urlgtf=
    "ftp://ftp.ensembl.org/pub/release-87/gtf/homo_sapiens/Homo_sapiens.GRCh38.87.gtf.gz",
    threads=24)

    #running rsemstar index for mouse
    rsemstarIndex(group="docker",genome.folder="/data/scratch/mm10star",
    ensembl.urlgenome="ftp://ftp.ensembl.org/pub/release-87/fasta/mus_musculus/dna/Mus_musculus.GRCm38.dna.toplevel.fa.gz",
    ensembl.urlgtf="ftp://ftp.ensembl.org/pub/release-87/gtf/mus_musculus/Mus_musculus.GRCm38.87.gtf.gz",
    threads=24)

}
}
\author{
Raffaele Calogero
}
