% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsemstarUcscIndex.R
\name{rsemstarUscsIndex}
\alias{rsemstarUscsIndex}
\title{Generating rsem-star genome index}
\usage{
rsemstarUscsIndex(
  group = c("sudo", "docker"),
  genome.folder = getwd(),
  uscs.urlgenome = NULL,
  uscs.gtf = NULL,
  uscs.urlknownIsoforms = NULL,
  uscs.urlknownToLocusLink = NULL,
  threads = 1
)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{genome.folder, }{a character string indicating the folder where the indexed reference genome will be located}

\item{uscs.urlgenome, }{a character string indicating the URL from uscs download web page for the unmasked genome sequence of interest}

\item{uscs.gtf, }{a character string indicating the path of the GTF file for genome of interest}

\item{uscs.urlknownIsoforms, }{a character string indicating the URL from uscs download web page for the knowisoforms file for genome of interest}

\item{uscs.urlknownToLocusLink, }{a character string indicating the URL from uscs download web page for the knownToLocusLink file for genome of interest}

\item{threads, }{a number indicating the number of cores to be used from the application}
}
\value{
The indexed genome reference sequence
}
\description{
This function executes the docker container rsem-star1 where RSEM and STAR are installed. The index is created using ENSEMBL genome fasta file. User needs to provide the URL for ENSEMBL genome located in the ENSEMBL ftp
}
\examples{
\dontrun{
    #running rsemstar index for human
    rsemstarUscsIndex(group="sudo",genome.folder="/data/scratch/hg19UCSCstar",
    uscs.urlgenome=
    "http://hgdownload.soe.ucsc.edu/goldenPath/hg19/bigZips/chromFa.tar.gz",
    uscs.gtf=
    "/Users/raffaelecalogero/Desktop/hg19_ucsc.gtf.gz",
    uscs.urlknownIsoforms=
    "http://hgdownload.soe.ucsc.edu/goldenPath/hg19/database/knownIsoforms.txt.gz",
    uscs.urlknownToLocusLink=
    "http://hgdownload.soe.ucsc.edu/goldenPath/hg19/database/knownToLocusLink.txt.gz",
    threads=24)

}
}
\author{
Raffaele Calogero
}
