% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmonAnnotation.R
\name{salmonAnnotation}
\alias{salmonAnnotation}
\title{A function to annotate salmon output}
\usage{
salmonAnnotation(group = c("sudo", "docker"), fastq.folder, index.folder)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{fastq.folder, }{a character string indicating the folder where input data are located and where output will be written, it should contain quant.sf file generated by salmonCounts}

\item{index.folder, }{a character string indicating the folder where Salmon transcriptome index was created with salmonIndex.}
}
\description{
This function executes a docker that convert the transcripts count file generated by Salmon quasi-alignment in the same format of isoforms.result of  RSEM
}
\examples{
\dontrun{
system("wget http://130.192.119.59/public/test_R1.fastq.gz")
system("wget http://130.192.119.59/public/test_R2.fastq.gz")
library(docker4seq)
#running salmonCounts
salmonCounts(group="docker", scratch.folder="/scratch/users/rcaloger/", 
        fastq.folder=getwd(), index.folder="/archive/home/rcaloger/data/seqbox/salmonIndex.R/transcripts_index", 
        threads=24, seq.type="pe")

#converting in a format identical to rsem isoform.results
salmonAnnotation(group="docker", fastq.folder=getwd(), 
    index.folder="/archive/home/rcaloger/data/seqbox/salmonIndex.R/transcripts_index")
}     
    
}
\author{
Raffaele Calogero, raffaele.calogero [at] unito [dot] it, Bioinformatics and Genomics unit, University of Torino Italy
}
