% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salmonCounts.R
\name{salmonCounts}
\alias{salmonCounts}
\title{A function to handle a salmon docker container}
\usage{
salmonCounts(group = c("sudo", "docker"), scratch.folder, fastq.folder,
  index.folder, threads = 8, seq.type = c("se", "pe"))
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{fastq.folder, }{a character string indicating the folder where input data are located and where output will be written}

\item{index.folder, }{a character string indicating the folder where transcriptome index was created with salmonIndex.}

\item{threads, }{a number indicating the number of cores to be used from the application}

\item{seq.type, }{a character string indicating the type of reads to be generated by the sequencer. Two options: \code{"se"} or \code{"pe"} respectively for single end and pair end sequencing. Strandness is inferred by salmon.}
}
\description{
This function executes a docker that produces as output the transcripts count file generated by Salmon quasi-alignment
}
\examples{
\dontrun{
    #running salmonCounts
    salmonCounts(group="docker", scratch.folder="/Users/raffaelecalogero/Desktop/scratch", 
    fastq.folder=getwd(), index.folder, threads=8, seq.type="pe")
}

}
\author{
Raffaele Calogero, raffaele.calogero [at] unito [dot] it, Bioinformatics and Genomics unit, University of Torino Italy
}
