% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSizeAndPower.R
\name{sampleSize}
\alias{sampleSize}
\title{A wrapper function to sample_size_distribution from RnaSeqSampleSize}
\usage{
sampleSize(filename, power = 0.8, FDR = 0.1, genes4dispersion = 200,
  log2fold.change = 1)
}
\arguments{
\item{filename, }{a character string indicating the name of the count table file}

\item{power, }{oower to detect prognostic genes}

\item{FDR, }{FDR level}

\item{genes4dispersion, }{number of genes used in estimation of read counts and dispersion distribution.}

\item{log2fold.change, }{a number indicating the minimum log2 fold changes for prognostic genes between two groups.}
}
\value{
a string with the requested informations
}
\description{
This function executes sample_size_distribution to identify the number of samples x group needed to obtain a specific statistical power
}
\examples{
\dontrun{
 sampleSize("_counts.txt", power=0.80, FDR=0.1, genes4dispersion=200, log2fold.change=1)

}
}
