% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samplePca.R
\name{samplesPca}
\alias{samplesPca}
\title{Sample PCA on raw counts}
\usage{
samplesPca(data, groups = NULL, output.file = NULL, label, label.size)
}
\arguments{
\item{data, }{the counts dataframe generated by samples2experiemtn function}

\item{groups, }{a character string indicating association between samples and experimental groups}

\item{output.file, }{a tab delimited file in which the first column are names from counts object from experiment.table.Rda file generated with samples2experiment function. In the pdf will be also rpesent the istigram fo the PCA variance dispersion.}

\item{label, }{a boolean value, TRUE or FALSE to ad the names of the samples in the plot}

\item{label.size, }{an integer value indicating the size of label}
}
\value{
a pdf file with the PCA plot
}
\description{
This function runs PCA via ggfortify and ggplot2 on the output of samples2experiments function
}
\examples{
\dontrun{
   load("test_experiment.tables.Rda")
   samplesPca(data=counts, groups=c("d","eb","wb","ws"),
   output.file="test.pdf", label = FALSE, label.size = 3)
}
}

