% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{skeleton}
\alias{skeleton}
\title{A skeleton function to handle a docker containier}
\usage{
skeleton(group = c("sudo", "docker"), scratch.folder, data.folder)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{data.folder, }{a character string indicating the folder where input data are located and where output will be written}
}
\description{
This function executes a ubuntu docker that produces as output helloworld.txt file
}
\examples{
\dontrun{
    #running skeleton
    skeleton(group="docker", scratch.folder="/Users/raffaelecalogero/Desktop/scratch", 
    data.folder=getwd())
}

}
\author{
Name Family name, myemail [at] somewhere [dot] org, Affiliation
}
