% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewer.R
\name{skewer}
\alias{skewer}
\title{Running skewer, an adapter trimmer application, Jiang et al BMC Bioinformatics201415:182}
\usage{
skewer(
  group = c("sudo", "docker"),
  fastq.folder = getwd(),
  scratch.folder = "/data/scratch",
  adapter5,
  adapter3,
  seq.type = c("se", "pe"),
  threads = 1,
  min.length = 18
)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{fastq.folder, }{a character string indicating where gzip fastq files are located}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{adapter5, }{a character string indicating the fwd adapter}

\item{adapter3, }{a character string indicating the rev adapter}

\item{seq.type, }{a character string indicating the type of reads to be trimmed. Two options: \code{"se"} or \code{"pe"} respectively for single end and pair end sequencing.}

\item{threads, }{a number indicating the number of cores to be used from the application}

\item{min.length, }{a number indicating minimal length required to return a trimmed read}
}
\value{
One or two gzip fastq files ending with trimmed-pair1.fastq.gz and trimmed-pair1.fastq.gz, a log file of the trimming with the extensione trimmed.log, run.info file descring the analysis steps done by the docker. The latter file is useful to understand where the docker stop in case of unexpected end
}
\description{
This function executes the docker container skewer1 to remove sequencing adapters from RNAseq reads
}
\examples{
\dontrun{
    system("wget http://130.192.119.59/public/test_R1.fastq.gz")
    system("wget http://130.192.119.59/public/test_R2.fastq.gz")
    skewer(group="docker",fastq.folder=getwd(), scratch.folder="/data/scratch",
    adapter5="AGATCGGAAGAGCACACGTCTGAACTCCAGTCA",
    adapter3="AGATCGGAAGAGCGTCGTGTAGGGAAAGAGTGT",
    seq.type="pe", threads=10,  min.length=40)
}
}
\author{
Raffaele Calogero
}
