% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sncRNA.R
\name{sncRNA}
\alias{sncRNA}
\title{Running small RNA-seq single-end reads alignment and quantification using BWA and custom scripts}
\usage{
sncRNA(group = c("sudo", "docker"), fastq.folder = getwd(), scratch.folder,
  ref.folder, nh = T, threads = 1, fastq.id, sample.id, g.id, refh.id,
  refnh.id)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{fastq.folder, }{a character string indicating where trimmed fastq files are located}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{ref.folder, }{a character string indicating the folder where the indexed reference genome for bwa is located}

\item{nh, }{a booleal value indicating if the analysis of non human small RNAs should be performed}

\item{threads, }{a number indicating the number of cores to be used from the application}

\item{fastq.id, }{a character string indicating the name of the fastq file to analyse}

\item{sample.id, }{a character string indicating the unique identifier of the dataset to analyse.}

\item{g.id, }{character string indicating the name human reference genome}

\item{refh.id, }{character string indicating the name human reference small RNA gene annotations}

\item{refnh.id, }{list of character strings indicating the name of references of non human small RNA annotations}
}
\value{
read count files:
}
\description{
This function executes the docker container where BWA is installed. BWA is a read alignment package that efficiently align short sequencing reads against a large reference sequence. Alignment is performed against annotations of human small RNAs. Read count is performed by GenomicAlignments R package and custom Python and bash commands.
}
\examples{
\dontrun{
    #downloading fastq files
    system("wget http://130.192.119.59/public/test_R1.fastq.gz")
    #running sncRNA quantification pipeline
    sncRNA(group="docker",fastq.folder=getwd(), scratch.folder="/data/scratch",
    ref.folder="/data/ref", nh=TRUE, fastq.id=test_R1.fastq.gz,
    threads=24, sample.id="test", g.id=hg38.fa, refh.id = small_ncRNA.fa, refnh.id = c(virus.fa, bacteria.fa))

}
}
\author{
Giulio Ferrero
}
