% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snvPreprocessing.R
\name{snvPreprocessing}
\alias{snvPreprocessing}
\title{Running indel relignment and quality recalibration on bam generated by bwa function NOT READY TO GO ON STABLE}
\usage{
snvPreprocessing(
  group = c("sudo", "docker"),
  bam.folder = getwd(),
  scratch.folder = "/data/scratch",
  genome.folder,
  threads = 8,
  gatk.file
)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{bam.folder, }{a character string indicating where bam file generated by bwa is located}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{genome.folder, }{a character string indicating the folder where the index reference genome for bwa is located}

\item{threads, }{a number indicating the number of cores to be used by the application}

\item{gatk.file, }{a character string for GenomeAnalysisTK-X.Y.tar.bz2, this file should be located in the bam.folder}
}
\value{
bam and bai after indel relignment and quality recalibration.
}
\description{
This function executes the docker container snvPreprocessing. IMPORTANT: the working folder require the presence of GenomeAnalysisTK.tar.bz2 files
}
\examples{
\dontrun{
    #downloading examples 1 million reads of mcf7 exome mixed with 1 million of mouse derived by human exome capturing
    system("wget http://130.192.119.59/public/hs1m_mm1m_R1.fastq.gz")
    system("wget http://130.192.119.59/public/hs1m_mm1m_R2.fastq.gz")
    
    #required for bwa 61Gb At the present time this is required to run mutect1
    system("wget http://130.192.119.59/public/hg19_exome.tar.gz")
    #this genome index is prepared to be used for indel realignment and quality recalibration
    
    #running snvPreprocessing
    snvPreprocessing(group="docker",bam.folder=getwd(), 
               scratch.folder="/data/scratch",
               genome.folder="/data/scratch/hg19_exome", 
               threads=24, gatk.file="GenomeAnalysisTK-3.7.tar.bz2")

}
}
\author{
Raffaele Calogero
}
