% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starChimeric.R
\name{starChimeric}
\alias{starChimeric}
\title{Running Star to detect chimeric transcripts on paired-end sequences}
\usage{
starChimeric(group = c("sudo", "docker"), fastq.folder = getwd(),
  scratch.folder = "/data/scratch", genome.folder, threads = 1,
  chimSegmentMin = 20)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{fastq.folder, }{a character string indicating where gzip fastq files are located}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{genome.folder, }{a character string indicating the folder where the indexed reference genome for STAR is located.}

\item{threads, }{a number indicating the number of cores to be used from the application}

\item{chimSegmentMin, }{is a positive value indicating the minimal lenght of the overlap of a read to the chimeric element}
}
\value{
three files: dedup_reads.bam, which is sorted and duplicates marked bam file, dedup_reads.bai, which is the index of the dedup_reads.bam, and dedup_reads.stats, which provides mapping statistics
}
\description{
This function executes STAR to detect chimeric transcripts
}
\examples{
\dontrun{
    #downloading fastq files
    system("wget http://130.192.119.59/public/test_R1.fastq.gz")
    system("wget http://130.192.119.59/public/test_R2.fastq.gz")
    #running star2step nostrand pe
    starChimeric(group="docker",fastq.folder=getwd(), scratch.folder="/data/scratch",
    genome.folder="/data/scratch/hg38star", threads=8, chimSegmentMin=20)

}
}
\author{
Raffaele Calogero, raffaele.calogero [at] unito [dot] it, Bioinformatics and Genomics unit, University of Torino Italy
}
