% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starchipCircle.R
\name{starchipCircle}
\alias{starchipCircle}
\title{Running starchip to detect circular RNAs on paired-end sequences}
\usage{
starchipCircle(group = c("sudo", "docker"), genome.folder, samples.folder,
  reads.cutoff = 5, min.subject.limit = 10, threads = 8,
  do.splice = c("True", "False"), cpm.cutoff = 0, subjectCPM.cutoff = 0,
  annotation = c("true", "false"))
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{genome.folder, }{a character string indicating the folder where the indexed reference genome for STAR is located.}

\item{samples.folder, }{the folder where are located all the folders of the samples processed with starChimeric}

\item{reads.cutoff, }{Integer. Minimum number of reads crossing the circular RNA backsplice required.}

\item{min.subject.limit, }{Integer. Minimum number of individuals with readsCutoff reads required to carry forward a cRNA for analysis}

\item{threads, }{Integer. Number of threads to use}

\item{do.splice, }{true false. The splices within the cRNA be detected and reported. Linear splices are searched within each cRNA in each individual. Any linear splice with >= 60\% of the read count of the cRNA is considered a splice within the cRNA. Two files are then created, .consensus with most common splice pattern, and .allvariants with all reported splice patterns.}

\item{cpm.cutoff, }{Float. Reads counts are loaded into R and log2(CountsPerMillion) is calculated using the limma package. With cpmCutoff > 0, cRNA with log2(CPM) below this value will be filtered from this analysis}

\item{subjectCPM.cutoff, }{Integer. See above. This value is the lower limit for number of individuals required to have the cRNA expressed at a value higher than cpmCutoff.}

\item{annotation, }{rue/false. cRNA are provided with gene annotations}
}
\value{
1. Count matrixes : raw cRNA backsplice counts: circRNA.cutoff[readthreshold]reads.[subjectthreshold]ind.countmatrix log2CPM of above: norm_log2_counts_circRNA.[readthreshold]reads.[subjectthreshold]ind.0cpm_0samples.txt Maximum Linear Splices at Circular Loci: rawdata/linear.[readthreshold]reads.[subjectthreshold]ind.sjmax 2. Info about each circRNA:  Consensus Information about Internal Splicing: Circs[reads].[subjects].spliced.consensus Complete Gene Annotation: circRNA.[readthreshold]reads.[subjectthreshold]ind.annotated Consise Gene Annotation + Splice Type:  circRNA.[readthreshold]reads.[subjectthreshold]ind.genes 3. Images: PCA plots: circRNA.[readthreshold]reads.[subjectthreshold]ind.0cpm_0samples_variance_PCA.pdf Heatmap: circRNA.[readthreshold]reads.[subjectthreshold]ind.heatmap.pdf
}
\description{
This function execute starchip on a set of folders containing the output of starChimeric. It requires a specific bed generated with starChipIndex in teh genome folder used by starChimeric
}
\examples{
\dontrun{
    #downloading fastq files
    system("wget http://130.192.119.59/public/test_R1.fastq.gz")
    system("wget http://130.192.119.59/public/test_R2.fastq.gz")
    #running star2step nostrand pe
    starchipCircle(group="docker", genome.folder="/data/genomes/hg38star", samples.folder=getwd(),
                       reads.cutoff=5, min.subject.limit=10, threads=8,
                       do.splice = "True"), cpm.cutoff=0,
}
}
\author{
Raffaele Calogero, raffaele.calogero [at] unito [dot] it, Bioinformatics and Genomics unit, University of Torino Italy
}
