% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starChipIndex.R
\name{starChipIndex}
\alias{starChipIndex}
\title{Preparing the GTF file required by starchip to detect circular RNAs on paired-end sequences}
\usage{
starChipIndex(group = c("sudo", "docker"), genome.folder = getwd())
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{genome.folder, }{a character string indicating where gzip fastq files are located}
}
\value{
three files: dedup_reads.bam, which is sorted and duplicates marked bam file, dedup_reads.bai, which is the index of the dedup_reads.bam, and dedup_reads.stats, which provides mapping statistics
}
\description{
This function prepare the genome GTF for starchip, the GTF must be the one used by starChimeric
}
\examples{
\dontrun{
    #downloading fastq files
    system("wget http://130.192.119.59/public/test_R1.fastq.gz")
    system("wget http://130.192.119.59/public/test_R2.fastq.gz")
    #running star2step nostrand pe
    starChipIndex(group="docker", genome.folder="/data/scratch/hg38star")
}
}
\author{
Raffaele Calogero, raffaele.calogero [at] unito [dot] it, Bioinformatics and Genomics unit, University of Torino Italy
}
