% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testSeqBox.R
\name{testSeqbox}
\alias{testSeqbox}
\title{Testing the SeqBox basic installation}
\usage{
testSeqbox()
}
\value{
testSeqBox.out file containing the results of the tests
}
\description{
This function executes a set of scripts to check if SeqBox installation is OK. It requests the existence of the following folders: /data/genomes, /data/scratch, /data/tests. /data/genomes folder must contain the following folders: hg38star  mm10bwa  mm10star
}
\author{
Raffaele Calogero
}
