% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimFasta.R
\name{trimFasta}
\alias{trimFasta}
\title{A function to execute seqtk trimming}
\usage{
trimFasta(
  group = c("sudo", "docker"),
  scratch.folder,
  data.folder,
  min.length = 300
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{data.folder, }{a character string indicating the folder where input data are located and where output will be written}

\item{min.length, }{min nucleotide lenght of each fasta seq}
}
\value{
Returns a fasta file colled trimmed.fasta.gz
}
\description{
This function executes a ubuntu docker that remove reads shorter of a specific threshold
}
\examples{
\dontrun{
    #running fastq2fasta
    trimfasta(group="docker", scratch.folder="/data/scratch", data.folder=getwd(), min.length=300)
}

}
\author{
Raffaele A Calogero, raffaele.calogero [at] unito [dot] it, University of Torino. Italy
}
