% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xenome.R
\name{xenome}
\alias{xenome}
\title{Running xenome, https://github.com/data61/gossamer/}
\usage{
xenome(
  group = c("sudo", "docker"),
  fastq.folder = getwd(),
  scratch.folder = "/data/scratch",
  xenome.folder,
  seq.type = "pe",
  threads = 1
)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{fastq.folder, }{a character string indicating where gzip fastq files are located}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{xenome.folder, }{a character string indicating the folder where the indexed reference genomes generated by xenome  are locates}

\item{seq.type, }{a character string indicating the type of reads to be trimmed. Two options: \code{"se"} or \code{"pe"} respectively for single end and pair end sequencing}

\item{threads, }{a number indicating the number of cores to be used from the application}
}
\value{
ambiguous, both, neither, hs and mm fastq.gz files. xeno_hs_R1.fastq.gz and xeno_hs_R2.fastq.gz are fastq file free of mouse reads and are used for further analysis.
}
\description{
This function executes the docker container bwa1 where BWA is installed BWA is a read alignment package that efficiently align short sequencing reads against a large reference sequence This aligner provides optimal results with DNA-seq data
}
\examples{
\dontrun{
    #downloading examples 1 million reads of mcf7 exome mixed with 1 million of mouse derived by human exome capturing
    system("wget http://130.192.119.59/public/hs1m_mm1m_R1.fastq.gz")
    system("wget http://130.192.119.59/public/hs1m_mm1m_R2.fastq.gz")
    #running xenome
    xenome(group="docker",fastq.folder=getwd(), scratch.folder="/data/scratch",
    xenome.folder="/data/scratch/hg19.mm10", seq.type="pe",
    threads=24)

}
}
\author{
Raffaele Calogero
}
