% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xenomeIndex.R
\name{xenomeIndex}
\alias{xenomeIndex}
\title{Generating xenome genome indexes}
\usage{
xenomeIndex(
  group = c("sudo", "docker"),
  xenome.folder = getwd(),
  hg.urlgenome = NULL,
  mm.urlgenome = NULL,
  threads = 8
)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{xenome.folder, }{a character string indicating the folder where the indexed reference genomes for xenome will be located}

\item{hg.urlgenome, }{a character string indicating the URL from uscs download web page for the unmasked human genome sequence of interest}

\item{mm.urlgenome, }{a character string indicating the URL from uscs download web page for the unmasked mouse genome sequence of interest}

\item{threads, }{an integer indicating how many threads are used by xenome}
}
\value{
The indexed xenome genomes references
}
\description{
This function executes the docker container xenome.2017.01 where xenome is installed.
}
\examples{
\dontrun{
    #running xenome index
    xenomeIndex(group="docker",xenome.folder="/data/scratch/test", hg.urlgenome=
    "http://hgdownload.soe.ucsc.edu/goldenPath/hg19/bigZips/chromFa.tar.gz",
    mm.urlgenome="http://hgdownload.cse.ucsc.edu/goldenPath/mm10/bigZips/chromFa.tar.gz")

}
}
\author{
Raffaele Calogero
}
