% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anovaLike.R
\name{anovaLike}
\alias{anovaLike}
\title{A function allowing the identification of differentially expressed genes if multiple groups are provided.}
\usage{
anovaLike(
  group = c("sudo", "docker"),
  file,
  logFC.threshold = 1,
  FDR.threshold,
  logCPM.threshold = 4,
  plot = c(TRUE, FALSE)
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{file, }{a character string indicating the path of the file, with counts.table name and extension included}

\item{logFC.threshold, }{minimal logFC present in at least one of the comparisons with respect to reference covariate}

\item{FDR.threshold, }{minimal FDR present in at least one of the comparisons with respect to reference covariate}

\item{logCPM.threshold, }{minimal average abundance}

\item{plot, }{TRUE if differentially expressed genes are represented in a plot.}
}
\description{
This function executes in a docker edgeR for the identification of differentially expressed genes in bulk RNAseq. IMPORTANT the filename shoould not have any '.' in the name unless .txt
}
\examples{
\dontrun{
    #running deDetection
    anovaLike(group="docker", file=paste(getwd(),"annotated_lorenz_buettner_counts_noSymb.txt", sep="/"),
       logFC.threshold=1, FDR.threshold=0.05, logCPM.threshold=4)
}

}
\author{
Raffaele Calogero, raffaele.calogero [at] unito [dot] it, University of Torino, Italy
}
