% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bwaIndex.R
\name{bwaIndex}
\alias{bwaIndex}
\title{Generating bwa genome index}
\usage{
bwaIndex(group = c("sudo", "docker"), genome.folder = getwd(), genome.fasta,
  genome.url = NULL, dbsnp.file = NULL, g1000.file = NULL, gatk = FALSE,
  download.genome = FALSE)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{genome.folder, }{a character string indicating the folder where the indexed reference genome for bwa will be located}

\item{genome.fasta, }{a character string indicating the local path to the genome fasta file to index}

\item{genome.url, }{a character string indicating the URL from download web page for the genome sequence of interest}

\item{dbsnp.file, }{a character string indicating the name of dbSNP vcf located in the genome folder. The dbSNP vcf, dbsnp_138.b37.vcf.gz and dbsnp_138.hg19.vcf.idx.gz, can be downloaded from ftp://gsapubftp-anonymous@ftp.broadinstitute.org/bundle/b37}

\item{g1000.file, }{a character string indicating the name of 1000 genome vcf located in the genome folder. The 1000 genomes vcf, Mills_and_1000G_gold_standard.indels.b37.vcf.gz and Mills_and_1000G_gold_standard.indels.hg19.sites.vcf.idx.gz, can be downloaded from ftp://gsapubftp-anonymous@ftp.broadinstitute.org/bundle/b37/}

\item{gatk, }{a boolean TRUE and FALSE that indicate if the index will be used for GATK analysis}

\item{download.genome, }{a boolean TRUE and FALSE that indicate if the genome file must but be downloaded}
}
\value{
The indexed bwa genome reference sequence
}
\description{
This function executes the docker container bwa1 where BWA is installed. Optionally, the index can be created also for GATK bundle data genome fasta file.
}
\examples{
\dontrun{

    #running bwa index
    bwaIndex(group="sudo",genome.folder="/data/genomes/mm10bwa", genome.fasta,
    gatk=FALSE, download.genome=FALSE)
    #running bwa index for gatk
    bwaIndex(group="sudo",genome.folder="/data/genomes/hg19_bwa", genome.url=
    "http://hgdownload.soe.ucsc.edu/goldenPath/hg19/bigZips/chromFa.tar.gz",
    dbsnp.file="dbsnp_138.hg19.vcf.gz", g1000.file="Mills_and_1000G_gold_standard.indels.hg19.sites.vcf.gz",
    gatk=TRUE, download.genome=TRUE)


}
}
\author{
Giulio Ferrero
}
