% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casc.impute.R
\name{cascImpute}
\alias{cascImpute}
\title{Imputing dropouts}
\usage{
cascImpute(group = c("sudo", "docker"), data.folder, counts.matrix, drop.thre,
  cores, refining = FALSE)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{data.folder, }{a character string indicating the folder where tab separated file of cells counts is saved}

\item{counts.matrix, }{a character string indicating the  tab separated file of cells  counts}

\item{drop.thre, }{A number between 0 and 1, specifying the threshold to determine dropout values}

\item{cores, }{a integer specifying the number of cores used for parallel computation.}

\item{refining, }{boolean FALSE, TRUE, if TRUE it execute again scImpute but setting a different drop.thre. Rerunning to change drop.thre hase to be done in the same folder where the first run was done.}
}
\value{
A matrix file with imputed data.
}
\description{
This function executes a wrapper for scImpute: Accurate And Robust Imputation For Single Cell RNA-Seq Data
}
\examples{
\dontrun{
    #downloading fastq files
    system("wget http://130.192.119.59/public/singlecells_counts.txt.gz")
    system("gzip -d singlecells_counts.txt.gz")
    cascImpute(group="docker", data.folder=getwd(), counts.matrix="singlecells_counts.txt", drop.thre=0.5, cores=8, refining=FALSE)
}

}
\author{
Raffaele Calogero, Luca Alessandri
}
