% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circrnaBSJunctions.R
\name{circrnaBSJunctions}
\alias{circrnaBSJunctions}
\title{Running CircHunter circRNA backsplicing sequence reconstruction module}
\usage{
circrnaBSJunctions(
  group = c("sudo", "docker"),
  scratch.folder,
  circrna.data,
  exon.data,
  assembly = "hg19"
)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{circrna.data, }{string indicating the path to the list of circRNAs}

\item{exon.data, }{string indicating the path to the exon annotation file}

\item{assembly, }{string indicating the reference human genome assembly. Compatible assemblies: hg19 (default), hg18, hg38, mm9, mm10, rn6, dm6, ce11.}
}
\value{
A fasta file reporting the sequence of circRNA back-splice junctions
}
\description{
This function executes the docker container circhunter by running the circRNA back-splicing sequence reconstruction module of CircHunter starting from a set of circRNAs. For CircHunter algorithm detail please refer to: https://github.com/carlo-deintinis/circhunter/tree/master/CircHunter.
}
\examples{
\dontrun{

    #retrieve the example data
    system("wget https://github.com/carlo-deintinis/circhunter/archive/master.zip")
    system("unzip master.zip")

    #running the circrnaBSJunctions function
    circrnaBSJunctions(group="docker", scratch.folder="/data/scratch", circrna.data=paste(getwd(),"/circhunter-master/CircHunter/toyexample/toy_circRNA", sep=""), exon.data=paste(getwd(),"/circhunter-master/CircHunter/toyexample/toy_genome", sep=""), assembly="hg19")

}
}
\author{
Nicola Licheri and Giulio Ferrero
}
