% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circrnaClassification.R
\name{circrnaClassification}
\alias{circrnaClassification}
\title{Running CircHunter circRNA classification module}
\usage{
circrnaClassification(
  group = c("sudo", "docker"),
  scratch.folder,
  circrna.data,
  exon.data,
  isoform.data,
  assembly = "hg19"
)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{circrna.data, }{string indicating the path to the list of circRNAs}

\item{exon.data, }{string indicating the path to the exon annotation file}

\item{isoform.data, }{string indicating the path to the isoform annotation file}

\item{assembly, }{string indicating the reference human genome assembly. Compatible assemblies: hg19 (default), hg18, hg38, mm9, mm10, rn6, dm6, ce11}
}
\value{
Two tab-delimited tables reporting the transcript- and gene-level classification of a list of circRNAs
}
\description{
This function executes the docker container circhunter by running the circRNA classification module of CircHunter starting from a set of circRNAs. For CircHunter algorithm detail please refer to: https://github.com/carlo-deintinis/circhunter/tree/master/CircHunter.
}
\examples{
\dontrun{

    #retrieve the example data
    system("wget https://github.com/carlo-deintinis/circhunter/archive/master.zip") #retrieve the example data
    system("unzip master.zip")

    #running the circrnaClassification function
    circrnaClassification(group="docker", scratch.folder="/data/scratch", circrna.data=paste(getwd(),"/circhunter-master/CircHunter/toyexample/toy_circRNA", sep=""), exon.data=paste(getwd(),"/circhunter-master/CircHunter/toyexample/toy_genome", sep=""), isoform.data=paste(getwd(),"/circhunter-master/CircHunter/toyexample/toy_isoformdata", sep=""), assembly="hg19")

}
}
\author{
Nicola Licheri and Giulio Ferrero
}
