% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circrnaPrepareFiles.R
\name{circrnaPrepareFiles}
\alias{circrnaPrepareFiles}
\title{Function to prepare the CircHunter reference annotations}
\usage{
circrnaPrepareFiles(
  group = c("sudo", "docker"),
  scratch.folder,
  data.folder,
  assembly = "hg19",
  version = NULL
)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{data.folder, }{a character string indicating the data folder where the output files will be saved}

\item{assembly, }{string indicating the reference human genome assembly. Compatible assemblies: hg19 (default), hg18, hg38, mm9, mm10, rn6, dm6, ce11}

\item{version, }{Ensembl database version used for the analysis. If no version number is provided, the last version is considered}
}
\value{
Two tab-delimited tables reporting the exons and the transcript isoforms annotations of an user-selected human genome assembly
}
\description{
This function executes the docker container circhunter by running the circRNA classification module of CircHunter starting from a set of circRNAs. For CircHunter algorithm detail please refer to: https://github.com/carlo-deintinis/circhunter/tree/master/CircHunter.
}
\examples{
\dontrun{
    #Download the example data
    system("wget https://github.com/carlo-deintinis/circhunter/archive/master.zip")
    system("unzip master.zip")

    #running the circrnaPrepareFiles function
    circrnaPrepareFiles(group="docker", scratch.folder="/data/scratch", data.folder="/data/output", assembly="hg19")

}
}
\author{
Nicola Licheri and Giulio Ferrero
}
