% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ciri2.R
\name{ciri2}
\alias{ciri2}
\title{Running CIRI v2 tool for circRNAs prediction}
\usage{
ciri2(
  group = c("sudo", "docker"),
  scratch.folder,
  sam.file,
  genome.file,
  annotation.file = NA,
  max.span = 2e+05,
  stringency.value = c("high", "low", "zero"),
  quality.threshold = 10,
  threads = 1
)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{sam.file, }{a character string indicating the path to the RNA-Seq alignment SAM file from BWA}

\item{genome.file, }{a character string indicating the path to the Fasta file of the reference genomic sequence (it should be the same reference indexed for the BWA alignment)}

\item{annotation.file, }{a character string indicating the path to the GTF/GFF file reporting the reference gene annotations}

\item{max.span, }{an integer reporting the maximum spanning distance of a circRNA (default = 200000 bp)}

\item{stringency.value, }{the selected stringency level of the analysis. Three possible options are available: "high" (high stringency, default), in which CIRI2 only provides circRNAs supported by more than 2 distinct Paired Chiastic Clipping (PCC) signals; "low" (low stringency), CIRI2 only provides circRNAs supported by more than 2 junction reads; "zero", CIRI2 provides all circRNAs regardless junction read counts or PCC signals}

\item{quality.threshold, }{integer indicating the threshold for mapping quality of each segment of junction reads (default=10)}

\item{threads, }{integer indicating the number of threads used for the analysis (default=1)}
}
\value{
The list of CIRI 2 predicted circRNAs
}
\description{
This function executes the docker container ciri2 where CIRI v2.0.6 is installed and it provides the list of circRNAs predicted from a RNA-Seq experiment. For CIRI 2 tool detail refer to: "Gao, Y., Zhang, J., & Zhao, F. (2017). Circular RNA identification based on multiple seed matching. Brief Bioinform. 2018 Sep 28;19(5):803-810."
}
\examples{
\dontrun{

    #retrieve the example data
    system("wget https://sourceforge.net/projects/ciri/files/CIRI2/CIRI_v2.0.6.zip") #retrieve the example data
    system("unzip CIRI_v2.0.6.zip")

    #running ciri2 function
    ciri2(group="docker", scratch.folder="/data/scratch", sam.file=paste(getwd(),"/CIRI_v2.0.6/data/sample.sam", sep=""), genome.file=paste(getwd(),"/CIRI_v2.0.6/data/chr1.fa", sep=""), annotation.file="", max.span=200000, stringency.value="high", quality.threshold=10, threads=1)
}

}
\author{
Nicola Licheri and Giulio Ferrero
}
