% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutadapt.R
\name{cutadapt}
\alias{cutadapt}
\title{Function to execute Cutadapt on RNA-Seq reads}
\usage{
cutadapt(
  group = c("sudo", "docker"),
  scratch.folder,
  data.folder,
  adapter.type = c("ILLUMINA", "NEB"),
  threads = 1
)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{data.folder, }{a character string indicating where fastq files are located}

\item{adapter.type, }{a character string. Two options: \code{"ILLUMINA"} or \code{"NEB"}, depending to which miRNA library prep was used: ILLUMINA or NEB}

\item{threads, }{a number indicating the number of cores to be used from the application}
}
\value{
The fastq files of the trimmed sequencing reads
}
\description{
This function executes the docker container  by running CutAdapt algorithm a the input RNA-Seq dataset to remove the sequencing adapters.
}
\examples{
\dontrun{
...
}
}
\author{
Nicola Licheri and Giulio Ferrero
}
