% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashclone.R
\name{hashclone}
\alias{hashclone}
\title{HashClone running A function to handle a hashclone docker container NOT READY for STABLE RELEASE, YET}
\usage{
hashclone(
  group = c("sudo", "docker"),
  scratch.folder,
  data.folder = getwd(),
  kmer,
  hash,
  coll,
  threshold = 0.1,
  type,
  spike = "null",
  input.files
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{data.folder, }{a character string indicating the folder where input data are located and where output will be written}

\item{kmer, }{an integer that define the size of the substrings (k-mer) encoded in the hash table (this must be a value between 1 and 32)}

\item{hash, }{a prime number indicating the size of the hash table. Increasing this value reduces the execution time but increases the memory utilization. Ideally, this value should be close to the number of different k-mers stored in the hash table;}

\item{coll, }{an integer that define the maximum number of different k-mers that the tool might need to store in the hash table.}

\item{threshold, }{(tau) this value is the threshold used to select significant k-mers. We suggest to set tau equal to 1}

\item{type, }{IGH (immunoglobulin heavy chain) or  IGK (immunoglobulin kappa locus)}

\item{spike, }{a character string indicating the path of the spike in file (if you don't want the spike in research, please set this parameter as 'null')}

\item{input.files, }{a character string indicating the path of the input files}
}
\description{
This function executes  HashClone algorithm developed to identify the set of clonality markers during the patient follow-up in order to quantify the minimal residual disease.
}
\examples{
\dontrun{
library(docker4seq)
downloadContainers(group="docker","docker.io/qbioturin/hashclone")
hashclone(group="docker",scratch.folder="/home/scratch_folder", data.folder="/home/output_folder", kmer=26, hash=10999997, coll=10999997, threshold=1, type= IGH, spike="/home/spike_in.fa", input.files=c('/home/input_file1.fastq', '/home/input_file2.fastq'))
}


}
\author{
Beccuti Marco, Greta Romano, Francesca Cordero, Raffaele Calogero, beccuti[at]di[dot]unito[dot]it, Computer Science Department Univ. of Turin.
}
