% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapBase.R
\name{heatmapBase}
\alias{heatmapBase}
\title{hfc, heatmap for clustering}
\usage{
heatmapBase(group = c("sudo", "docker"), scratch.folder, file,
  geneNameControl = 1, status = 0)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with counts.table name and extension included}

\item{geneNameControl, }{0 if the matrix has gene name without ENSEMBL code.1 if the gene names is formatted like this : ENSMUSG00000000001:Gnai3. If the gene names is only ensamble name you have to run SCannoByGtf before start using this script.}

\item{status, }{0 if is raw count, 1 otherwise}
}
\value{
A heatmap.
}
\description{
This function generate heatmap and other plot based on clustering and on a specific gene list
}
\examples{
\dontrun{
system("wget http://130.192.119.59/public/heatmap_test.zip")
system("unzip heatmap_test.zip")
setwd("heatmap_test")
heatmapBase(group="docker",scratch.folder="/data/scratch",file=paste(getwd(),"DEfiltered__log2TPM", sep="/"), geneNameControl=1, status=hcl1)
}
}
\author{
Luca Alessandri , alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
