% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapBase.R
\name{heatmapBase}
\alias{heatmapBase}
\title{hfc, heatmap for clustering}
\usage{
heatmapBase(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  status = 0,
  lower.range = -1,
  upper.range = 1
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{file, }{a character string indicating the path of the file, with counts.table name and extension included}

\item{status, }{0 if is raw count, 1 otherwise}

\item{lower.range, }{the lower range of signal in the heatmap.}

\item{upper.range, }{the upper range of signal in the heatmap.}
}
\value{
A heatmap.
}
\description{
This function generate heatmap and other plot based on clustering and on a specific gene list
}
\examples{
\dontrun{
system("wget http://130.192.119.59/public/heatmap_test.zip")
system("unzip heatmap_test.zip")
setwd("heatmap_test")
heatmapBase(group="docker",scratch.folder="/data/scratch",file=paste(getwd(),"DEfiltered__log2TPM.txt", sep="/"), status=1, lower.range=-1, upper.range=1)
}
}
\author{
Luca Alessandri , alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
