% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indropCounts.R
\name{indropCounts}
\alias{indropCounts}
\title{A function to handle a indrop V2 single cell data}
\usage{
indropCounts(group = c("sudo", "docker"), scratch.folder, fastq.folder,
  index.folder, sample.name, split.affixes, bowtie.index.prefix, M = 10,
  U = 2, D = 400, low.complexity.mask = c("False", "True"),
  umi.threshold = 5)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{fastq.folder, }{a character string indicating the folder where input data are located and where output will be written}

\item{index.folder, }{a character string indicating the folder where transcriptome index was created with salmonIndex.}

\item{sample.name, }{the name to be associated to the fastq files, e.g. C2 for C2_S2_L001_R1_001.fastq.gz, IMPORTANT input fastq should have the format SAMPLENAME_Sx_L00y_Rz_001.fastq.gz, where x is an integer, y is an integer, z is 1 or 2}

\item{split.affixes, }{the string separating SAMPLENAME for the Rz_001.fastq.gz}

\item{bowtie.index.prefix, }{the prefix name of the bowtie index}

\item{M, }{Ignore reads with more than M alignments, after filtering on distance from transcript end.}

\item{U, }{Ignore counts from UMI that should be split among more than U genes.}

\item{D, }{Maximal distance from transcript end, NOT INCLUDING THE POLYA TAIL.}

\item{low.complexity.mask, }{low complexity mask False or True}

\item{umi.threshold, }{the minimal number of UMI to consider a gene present}
}
\description{
This function executes a docker that produces as output the sinngle cell counts from V2 indrop single cell sequencing
}
\examples{
\dontrun{
system("wget 130.192.119.59/public/testMm_S0_L001_R1_001.fastq.gz")
system("wget 130.192.119.59/public/testMm_S0_L001_R2_001.fastq.gz")
library(docker4seq)
#running indropCounts
indropCounts(group="docker", scratch.folder="/data/scratch", fastq.folder=getwd(),
       index.folder="/data/genomes/mm10indrop", sample.name="testMm", split.affixes="S0_L001",
       bowtie.index.prefix="Mus_musculus.GRCm38.85.index", M=10, U=2, D=400, low.complexity.mask="False", umi.threshold=5)
}

}
\author{
Raffaele Calogero and Riccardo Panero, raffaele.calogero [at] unito [dot] it, Bioinformatics and Genomics unit, University of Torino Italy
}
