% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutationClustering.R
\name{permutationClustering}
\alias{permutationClustering}
\title{Permutations and Clustering}
\usage{
permutationClustering(group = c("sudo", "docker"), scratch.folder,
  data.folder, matrixName, nPerm, permAtTime, percent, range1, range2, format,
  separator, logTen, clustering)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{data.folder, }{a character string indicating the folder where input data are located and where output will be written}

\item{matrixName, }{counts table name. Matrix data file must be in data.folder. The file MUST contain RAW counts, without any modification, such as log transformation, normalizatio etc.}

\item{nPerm, }{number of permutations to perform the pValue to evaluate clustering}

\item{permAtTime, }{number of permutations that can be computes in parallel}

\item{percent, }{percentage of random cells that has to be removed in each permutation}

\item{range1, }{first number of cluster for k means algorithm}

\item{range2, }{last number of cluster for k means algorithm}

\item{format, }{count matrix format "csv", "txt"..}

\item{separator, }{separator used in count file, e.g. '\\t', ','}

\item{logTen, }{1 if the count matrix is already in log10, 0 otherwise}

\item{clustering, }{clustering method to use : "SIMLR" , "tsne", "griph"}
}
\value{
VioPlot of silhouette cells value for each number of cluster used,clusterP file with clustering results for each permutation, killedCell file with removed cells in each permutation, clustering.output a sommarize file with general information for each cells.
}
\description{
This function executes a ubuntu docker that produces a specific number of permutation to evaluate clustering.
}
\examples{
\dontrun{
 permutationClustering("sudo","/home/lucastormreig/CASC2.0/permutationClustering/scratch/","/home/lucastormreig/CASC2.0/permutationClustering/Data/","TOTAL",4,2,10,3,4,"csv",",",0,"SIMLR")# 
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
