% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaseqCounts.R
\name{rnaseqCounts}
\alias{rnaseqCounts}
\title{Running RNAseq counting workflow for a single sample}
\usage{
rnaseqCounts(
  group = "sudo",
  fastq.folder = getwd(),
  scratch.folder = "/data/scratch",
  threads = 4,
  adapter5,
  adapter3,
  seq.type = "pe",
  min.length = 40,
  genome.folder = "/data/genomes/hg38star",
  strandness = "none",
  save.bam = TRUE,
  org = "hg38",
  annotation.type = "gtfENSEMBL"
)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{fastq.folder, }{a character string indicating where gzip fastq files are located}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{threads, }{a number indicating the number of cores to be used from the application}

\item{adapter5, }{a character string indicating the fwd adapter}

\item{adapter3, }{a character string indicating the rev adapter}

\item{seq.type, }{a character string indicating the type of reads to be generated by the sequencer. Two options: \code{"se"} or \code{"pe"} respectively for single end and pair end sequencing.}

\item{min.length, }{a number indicating minimal length required to return a trimmed read
#params rsemstar}

\item{genome.folder, }{a character string indicating the folder where the indexed reference genome is located. IMPORTANT the present function only suport genomic indexes made using ensembl genom and the corresponding gtf}

\item{strandness, }{a character string indicating the type ofsequencing protocol used for the analysis. Three options: \code{"none"}, \code{"forward"}, \code{"reverse"} respectively for non strand selection, reverse for Illumina strandness protocols, reverse for ACCESS Illumina protocol}

\item{save.bam, }{a boolean value, TRUE or FALSE, to save also BAM files generated by STAR and RSEM
#params rsemanno}

\item{org, }{a character string indicating the genome assembly used for mapping and counting with \code{"rsemstar"} function only required for biocENSEMBL based annotation}

\item{annotation.type, }{a character string. Two options: \code{"biocENSEMBL"} or \code{"gtfENSEMBL"}. \code{"biocENSEMBL"} will annotate by Bioconductor only protein coding genes. \code{"gtfENSEMBL"} will annotate all RNAs described in \code{"annotation.type"}}
}
\value{
Returns the output of skewer, rsemstar, rsemannos' functions
}
\description{
This function executes a set of docker containers allowing the generation of gene and isoforms counts for a single sample.
#params skewer
}
\examples{
\dontrun{
system("wget http://130.192.119.59/public/test_R1.fastq.gz")
library(docker4seq)
rnaseqCounts(group="docker",fastq.folder=getwd(), scratch.folder="/data/scratch/",
            adapter5="AGATCGGAAGAGCACACGTCTGAACTCCAGTCA",
            adapter3="AGATCGGAAGAGCGTCGTGTAGGGAAAGAGTGT",
            seq.type="se", threads=24,  min.length=40,
            genome.folder="/data/genomes/hg38star", strandness="none", save.bam=FALSE,
            org="hg38", annotation.type="gtfENSEMBL")
}
}
\author{
Raffaele Calogero
}
