% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatingByGtf.R
\name{rsemannoByGtf}
\alias{rsemannoByGtf}
\title{Annotating RSEM gene.results using ENSEMBL gtf and refGenome CRAN package}
\usage{
rsemannoByGtf(group = "docker", rsem.folder = getwd(), genome.folder)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{rsem.folder, }{a character string indicating where gene.results file is located}

\item{genome.folder, }{a character string indicating the folder for the genome reference used for mapping and counting with \code{"rsemstar"} function. In this folder is present the GTF used for by RSEM}
}
\value{
one file: annotated_genes.results, which is the annotated version of gene.results.
}
\description{
This function executes the docker container annotate.1, where refGenome is used to annotated gene.results and isoforms.results outputs from RSEM using ENSEMBL GTF annotation
}
\examples{
\dontrun{
    #downloading fastq files
    system("wget http://130.192.119.59/public/test_R1.fastq.gz")
    system("wget http://130.192.119.59/public/test_R2.fastq.gz")
    library(docker4seq)
   #running rsemstar nostrand pe
   rsemstar(group="docker",fastq.folder=getwd(), scratch.folder="/data/scratch/",
         genome.folder="/data/genomes/hg38star/", seq.type="pe", strandness="none",
         threads=8, save.bam = FALSE)
    #running rsemannoByGtf
    rsemannoByGtf(group="docker", rsem.folder=getwd(), genome.folder="/data/scratch/hg38star")
}

}
\author{
Raffaele Calogero
}
