% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsemStar.R
\name{rsemstar}
\alias{rsemstar}
\title{Running RSEM, Li and Dewey BMC Bioinformatics 2011 12:323}
\usage{
rsemstar(
  group = c("sudo", "docker"),
  fastq.folder = getwd(),
  scratch.folder = "/data/scratch",
  genome.folder,
  seq.type = c("se", "pe"),
  strandness = c("none", "forward", "reverse"),
  threads = 1,
  save.bam = TRUE
)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{fastq.folder, }{a character string indicating where gzip fastq files are located}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{genome.folder, }{a character string indicating the folder where the indexed reference genome for STAR is located. IMPORTANT the present function only suport genomic indexes made using ensembl genom and the corresponding gtf}

\item{seq.type, }{a character string indicating the type of reads to be trimmed. Two options: \code{"se"} or \code{"pe"} respectively for single end and pair end sequencing}

\item{strandness, }{a character string indicating the type ofsequencing protocol used for the analysis. Three options: \code{"none"}, \code{"forward"}, \code{"reverse"} respectively for non strand selection, forward for Illumina strandness protocols, reverse for ACCESS Illumina protocol}

\item{threads, }{a number indicating the number of cores to be used from the application}

\item{save.bam, }{a boolean TRUE FALSE to decide if bam files are saved}
}
\value{
three files: dedup_reads.bam, which is sorted and duplicates marked bam file, dedup_reads.bai, which is the index of the dedup_reads.bam, and dedup_reads.stats, which provides mapping statistics
}
\description{
This function executes the docker container rsemstar1, where RSEM is used to calculate gene/isoforms counts using as mapper STAR, Dubin et al. Bioinformatics. 2013 Jan 1;29(1):15-21
}
\examples{
\dontrun{
    #downloading fastq files
system("wget http://130.192.119.59/public/test_R1.fastq.gz")
system("wget http://130.192.119.59/public/test_R2.fastq.gz")
library(docker4seq)
#running rsemstar nostrand pe
rsemstar(group="docker",fastq.folder=getwd(), scratch.folder="/data/scratch/",
         genome.folder="/data/genomes/hg38star/", seq.type="pe", strandness="none",
         threads=8, save.bam = FALSE)

}
}
\author{
Raffaele Calogero
}
