% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDocker.R
\name{runDocker}
\alias{runDocker}
\title{Run docker container}
\usage{
runDocker(group = "docker", params = NULL, DockerSwarm = FALSE)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{params, }{a character string containing all parameters needed to tun the docker container}

\item{DockerSwarm, }{a bolean value used to enable docker execution in swarm mode.}
}
\value{
0 if success, 1 if parameters are missing, 2 if the group is neither sudo or docker, 3 if docker execution fails.
}
\description{
This is an internal function executing a docker container. Not to be used by users.
}
\examples{
\dontrun{
    #running runDocker
     runDocker(group="docker", params=NULL)

}
}
\author{
Raffaele Calogero
}
