% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/samples2experiment.R
\name{sample2experiment}
\alias{sample2experiment}
\title{generating counts, FPKM and TPM tables from rnaseqCounts outuputs}
\usage{
sample2experiment(
  sample.folders,
  covariates,
  batch = NULL,
  bio.type = c("protein_coding", "unitary_pseudogene", "unprocessed_pseudogene",
    "processed_pseudogene", "transcribed_unprocessed_pseudogene", "processed_transcript",
    "antisense", "transcribed_unitary_pseudogene", "polymorphic_pseudogene", "lincRNA",
    "sense_intronic", "transcribed_processed_pseudogene", "sense_overlapping",
    "IG_V_pseudogene", "pseudogene", "TR_V_gene", "3prime_overlapping_ncRNA",
    "IG_V_gene", "bidirectional_promoter_lncRNA", "snRNA", "miRNA", "misc_RNA", "snoRNA",
    "rRNA", "IG_C_gene", "IG_J_gene",      "TR_J_gene", "TR_C_gene", "TR_V_pseudogene",
    "TR_J_pseudogene", "IG_D_gene", "ribozyme", "IG_C_pseudogene", "TR_D_gene", "TEC",
    "IG_J_pseudogene", "scRNA", "scaRNA", "vaultRNA", "sRNA", "macro_lncRNA",
    "non_coding", "IG_pseudogene"),
  output.prefix = "."
)
}
\arguments{
\item{sample.folders, }{a character string indicating the paths of rnaseqCouts output folders}

\item{covariates, }{a character string indicating the covariates associated to each sample. Covariates are required for differnetial expression analysis}

\item{batch, }{a character string indicating the batch associated to each sample}

\item{bio.type, }{a character string indicating the ensemb bio.type. Options: "protein_coding","unitary_pseudogene","unprocessed_pseudogene","processed_pseudogene", "transcribed_unprocessed_pseudogene","processed_transcript","antisense","transcribed_unitary_pseudogene","polymorphic_pseudogene","lincRNA","sense_intronic","transcribed_processed_pseudogene","sense_overlapping","IG_V_pseudogene","pseudogene","TR_V_gene","3prime_overlapping_ncRNA","IG_V_gene","bidirectional_promoter_lncRNA","snRNA","miRNA","misc_RNA","snoRNA","rRNA","IG_C_gene","IG_J_gene","TR_J_gene","TR_C_gene","TR_V_pseudogene","TR_J_pseudogene","IG_D_gene","ribozyme","IG_C_pseudogene","TR_D_gene","TEC","IG_J_pseudogene","scRNA","scaRNA","vaultRNA","sRNA","macro_lncRNA","non_coding","IG_pseudogene"}

\item{output.prefix, }{a character value indicating the output folder path}
}
\value{
Returns counts, fpkm, tpm data frames for gene and isoforms, save data frames in experiment.tables.Rda, in counts.txt, log2fpkm.txt and in log2TPM
}
\description{
This function generates counts, FPKM and TPM tables from rnaseqCounts outuputs.
}
\examples{
\dontrun{
  system("wget http://130.192.119.59/public/test.samples2experiment.zip")
  unzip("test.samples2experiment.zip")
  setwd("test.samples2experiment")
  library(docker4seq)
  sample2experiment(sample.folders=c("./e1g","./e2g","./e3g",
                 "./p1g", "./p2g", "./p3g"),
                 covariates=c("Cov.1","Cov.1","Cov.1",
                 "Cov.2","Cov.2","Cov.2"),
                 bio.type="protein_coding", output.prefix=".")
}
}
\author{
Raffaele Calogero
}
