% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sncRNA.R
\name{sncRNA}
\alias{sncRNA}
\title{Running small RNA-seq single-end reads alignment and quantification using BWA and custom scripts}
\usage{
sncRNA(group = c("sudo", "docker"), fastq.folder = getwd(), scratch.folder,
  mode, reference, threads = 1, mb.version = NULL, mb.species = NULL)
}
\arguments{
\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{fastq.folder, }{a character string indicating where trimmed fastq files are located}

\item{scratch.folder, }{a character string indicating the scratch folder where docker container will be mounted}

\item{mode, }{a character string indicating the required type of analysis. Compatible analyses mode are "miRNA" and "ncRNA". In "miRNA" analysis mode, the version ("mb.version" argument) and species prefix ("mb.species" argument) of miRBase are required. In "ncRNA" analysis mode, the version ("rc.version" argument) and species prefix ("rc.species" argument) of RNA Central are required. This mode require also a desidered maximum length of the studied RNA annotations ("length" argument).}

\item{reference, }{a character string indicating the path to the reference fasta file used to create the BWA index}

\item{threads, }{a number indicating the number of cores to be used from the application}

\item{mb.version, }{a character string indicating the required version of miRBase database. Visit ftp://mirbase.org/pub/mirbase/ to select the proper version id.}

\item{mb.species, }{a character string indicating the three-letter prefix of a species annotated in miRBase (e.g. "hsa" for human miRNAs). Please refer to http://www.mirbase.org/help/genome_summary.shtml to obtain the proper species prefix.}
}
\value{
read count files:
}
\description{
This function executes the docker container where BWA is installed. BWA is a read alignment package that efficiently align short sequencing reads against a large reference sequence. Alignment is performed against annotations of human small RNAs. Read count is performed by GenomicAlignments R package and custom Python and bash commands.
}
\examples{
\dontrun{
    #downloading fastq files
    system("wget http://130.192.119.59/public/test_R1.fastq.gz")
    #running miRNAs quantification pipeline
    sncRNA(group="docker", fastq.folder=getwd(), scratch.folder="/data/scratch", mode="miRNA", reference="/data/ref", threads=24, mb.version="22", mb.species="hsa")
    
    #running non miRNA ncRNAs quantification pipeline
    sncRNA(group="docker", fastq.folder=getwd(), scratch.folder="/data/scratch", mode="ncRNA", reference="/data/ref", threads=24)

}
}
\author{
Giulio Ferrero
}
