% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starchipCircle.R
\name{starchipCircle}
\alias{starchipCircle}
\title{Running starchip to detect circular RNAs on paired-end sequences}
\usage{
starchipCircle(group = c("sudo", "docker"), genome.folder, samples.folder)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{genome.folder, }{a character string indicating the folder where the indexed reference genome for STAR is located.}

\item{samples.folder, }{the folder where are located all the folders of the samples processed with starChimeric}
}
\value{
three files: dedup_reads.bam, which is sorted and duplicates marked bam file, dedup_reads.bai, which is the index of the dedup_reads.bam, and dedup_reads.stats, which provides mapping statistics
}
\description{
This function execute starchip on a set of folders containing the output of starChimeric. It requires a specific bed generated with starChipIndex in teh genome folder used by starChimeric
}
\examples{
\dontrun{
    #downloading fastq files
    system("wget http://130.192.119.59/public/test_R1.fastq.gz")
    system("wget http://130.192.119.59/public/test_R2.fastq.gz")
    #running star2step nostrand pe
    starchipCircle(group="docker", genome.folder="/data/genomes/hg38star", samples.folder=getwd())
}
}
\author{
Raffaele Calogero, raffaele.calogero [at] unito [dot] it, Bioinformatics and Genomics unit, University of Torino Italy
}
