% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starChipIndex.R
\name{starChipIndex}
\alias{starChipIndex}
\title{Preparing the bed file required by starchip to detect circular RNAs on paired-end sequences}
\usage{
starChipIndex(group = c("sudo", "docker"), genome.folder = getwd())
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{genome.folder, }{a character string indicating the folder where the indexed reference genome for STAR is located.}
}
\value{
the bed files of the reference genome for STARChip analysis
}
\description{
This function prepare the genome bed file for starchip, the GTF must be the one used by starChimeric
}
\examples{
\dontrun{
    starChipIndex(group="docker", genome.folder="/data/genomes/hg38star")
}
}
\author{
Raffaele Calogero, raffaele.calogero [at] unito [dot] it, Bioinformatics and Genomics unit, University of Torino Italy
}
