% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testSeqBox.R
\name{testSeqbox}
\alias{testSeqbox}
\title{Testing the SeqBox basic installation}
\usage{
testSeqbox()
}
\value{
testSeqBox.out file containing the results of the tests
}
\description{
This function executes a set of scripts to check if SeqBox installation is OK. It requests the existence of the following folders: /data/genomes, /data/scratch, /data/tests. /data/genomes folder must contain the following folders: hg38star  mm10bwa  mm10star
}
\examples{
\dontrun{
    #downloading in /data/genomes
    setwd("/data/genomes")
    system("wget http://130.192.119.59/public/hg38star.tar.gz")
    system("wget http://130.192.119.59/public/mm10star.tar.gz")
    system("wget http://130.192.119.59/public/mm10bwa.tar.gz")
    system("wget http://130.192.119.59/public/hg38salmon.tar.gz")
    system("gzip -d *.gz")
    system("tar xvf *.tar")
    setwd("/data/")
    system("wget http://130.192.119.59/public/tests.tar.gz")
    system("gzip -d tests.tar.gz")
    system("tar xvf tests.tar")
    system("rm tests.tar")
    #running test SeqBox
    library(docker4seq)
    testSeqbox()

}
}
\author{
Raffaele Calogero
}
