% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapperDeseq2.R
\name{wrapperDeseq2}
\alias{wrapperDeseq2}
\title{A wrapper function for deseq2 for two groups only}
\usage{
wrapperDeseq2(
  output.folder,
  group = c("sudo", "docker"),
  experiment.table,
  log2fc = 1,
  fdr = 0.1,
  ref.covar = "0",
  type = c("gene", "isoform", "mirna"),
  batch = FALSE
)
}
\arguments{
\item{output.folder, }{a character string indicating where the tables generated by samples2experiment are located and were results will be placed}

\item{group, }{a character string. Two options: \code{"sudo"} or \code{"docker"}, depending to which group the user belongs}

\item{experiment.table, }{a character string indicating the counts table generated with sample3experiment with addition of covariates}

\item{log2fc, }{log2fc threshold for differetially expressed genes}

\item{fdr, }{fdr threshold}

\item{ref.covar, }{covariate to be used as reference}

\item{type, }{character with three options: gene, isoform, mirna. if gene is used two files are generated for geneset enrichment, the filtered Gene symbols and the background that contains all gene simbols.}

\item{batch, }{logical FALSE, TRUE}
}
\value{
Returns a full table of differentially expressed genes (prefix DEfull), a filtered table of differentially expressed genes (prefix DEfiltered) and the normalized counts table (prefix normalized)
}
\description{
This function runs deseq2 on a table genrated with sample2experiment having the covariates added in the names of the columns, separated by the names with underscore.
}
\examples{
\dontrun{
    system("wget 130.192.119.59/public/test.analysis.zip")
    unzip("test.analysis.zip")
    setwd("test.analysis")
    library(docker4seq)
    wrapperDeseq2(output.folder=getwd(), group="docker", experiment.table="_counts.txt", log2fc=1,
    fdr=0.1, ref.covar="Cov.1", type="gene", batch=FALSE)

}
}
\author{
Raffaele Calogero
}
